@@ -1,151 +1,155 @@
-name: yessoft
-description: A new Flutter application.
+name: pasco_shipping
+description: A new Flutter project.
 
-publish_to: 'none'
+# The following line prevents the package from being accidentally published to
+# pub.dev using `pub publish`. This is preferred for private packages.
+publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
-version: 1.1.0
+# The following defines the version and build number for your application.
+# A version number is three numbers separated by dots, like 1.2.43
+# followed by an optional build number separated by a +.
+# Both the version and the builder number may be overridden in flutter
+# build by specifying --build-name and --build-number, respectively.
+# In Android, build-name is used as versionName while build-number used as versionCode.
+# Read more about Android versioning at https://developer.android.com/studio/publish/versioning
+# In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
+# Read more about iOS versioning at
+# https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
+version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^1.0.0
-  flutter_datetime_picker: ^1.4.0
 
-  # Localization Library
+
+  # The following adds the Cupertino Icons font to your application.
+  # Use with the CupertinoIcons class for iOS style icons.
+  cupertino_icons: ^1.0.2
+  # Custom SnackBar
+  another_flushbar: ^1.10.22
+  # Custom Toast
+  fluttertoast: ^8.0.7
+  # Local Notifications
+  flutter_local_notifications: ^6.0.0
+
   flutter_localizations:
     sdk: flutter
-
   # Fonts
-  google_fonts: ^1.1.1
-  font_awesome_flutter: ^8.11.0
-
-  # DI
-  provider: ^4.3.2+2
-
-  # Crypto for SHA1
-  crypto: ^2.1.5
-
-  # Dependency injection
-  inject:
-    git:
-      url: https://github.com/yes-soft-de/inject.dart.git
-      path: package/inject
-
-  # Camera
-  image_picker: ^0.6.7+14
-
-  # Video
-  video_player: ^1.0.0
-
-  # Paths
-  path_provider: ^1.6.24
-
-  # Shared Prefs
-  shared_preferences: ^0.5.12+4
-
-  # Open Url in Chrome
-  url_launcher: ^5.7.0
-
+  font_awesome_flutter: ^9.0.0
+  google_fonts: ^2.1.0
+  # Camera And Gallery
+  image_picker: ^0.8.0
+  # Shared Pref
+  shared_preferences: ^2.0.6
+  # NoSQL Database
+  hive: ^2.0.4
+  # Open Url On Browser
+  url_launcher: ^6.0.6
+  share: ^2.0.4
   # Network Agent
-  dio: ^3.0.10
-  dio_flutter_transformer: ^3.0.2
-  dio_http_cache: ^0.2.11
-  dio_firebase_performance: ^0.2.0
-
-  # Pair and Other Types
-  analyzer_plugin: ^0.3.0
-  timeago: ^2.0.29
-
-  # String validation
-  validators: ^2.0.1
-
+  dio: ^4.0.0
+  dio_firebase_performance: ^0.3.0
+  # Analyzer
+  analyzer_plugin: ^0.6.0
+  # Time Ago Generator
+  timeago: ^3.0.2
+  # String Validators
+  validators: ^3.0.0
   # Svg Loader
-  flutter_svg: ^0.19.1
-
-  # Streams and RX
-  rxdart: ^0.24.1
-
-  # ORM, Database SQL
-  sqfentity: ^1.4.0+3
-
-  # Firebase Libs
-  firebase_core: ^0.5.0
-  firebase_auth: ^0.18.4+1
-  firebase_analytics: ^6.3.0
-  firebase_remote_config: ^0.4.3
-  cloud_firestore: ^0.14.4
-  firebase_performance: ^0.4.0
-  firebase_messaging: ^7.0.0
-  firebase_crashlytics: ^0.2.4
-
-  # Google Sign In
-  google_sign_in: ^4.5.6
-  sign_in_with_apple: ^2.5.4
-
-  # Localization
-  gen_lang: 0.1.3
-
+  flutter_svg: ^0.22.0
+  # Stream And RX
+  rxdart: ^0.27.1
+  # Firebase
+  firebase_core: ^1.2.0
+  firebase_auth: ^1.2.0
+  firebase_analytics: ^8.1.0
+  firebase_remote_config: ^0.10.0
+  cloud_firestore: ^2.2.0
+  firebase_performance: ^0.7.0+3
+  firebase_messaging: ^10.0.0
+  firebase_crashlytics: ^2.0.4
+  # Sign In
+  google_sign_in: ^5.0.4
+  sign_in_with_apple: ^3.0.0
   # Internationalization
-  intl: ^0.16.1
-  intl_translation: ^0.17.10
-  intl_utils: ^1.9.0
-
-  # Toasts
-  fluttertoast: ^7.1.1
-
+  intl: ^0.17.0
+  intl_utils: ^2.3.0
   # UUID
-  uuid: ^2.2.2
-
-  # ProgressBar
-  step_progress_indicator: ^0.2.5+8
-
-  # Maps and Locations
-  flutter_map: ^0.10.1+1
-  latlong: ^0.6.1
-  location: ^3.1.0
-
-  # Deep Linking, for quick send
-  uni_links: ^0.4.0
-
-  # Charts and Diagrams
-  charts_flutter: ^0.9.0
-
-  # Credit Card and Payment
-  flutter_credit_card: ^0.1.3
-  stripe_payment : ^1.0.9
+  uuid: ^3.0.4
+  # Geo
+  latlong2: ^0.8.0
+  location: ^4.1.1
+  uni_links: ^0.5.1
+  # Payment Gates
+  flutter_credit_card: ^2.0.0
+  # Lottie And Flare Animation
+  lottie: ^1.0.1
+  flare_flutter: ^3.0.0
+  # DI
+  injectable: ^1.4.1
+  get_it: ^7.1.3
+  # Maps
+  flutter_map: ^0.13.0
 
 dev_dependencies:
+  # Test Libs
   flutter_test:
     sdk: flutter
+  flutter_gherkin: ^2.0.0
+  test: ^1.16.8
+  # Build Runner And DI Generator
+  build_runner: ^2.0.4
+  injectable_generator: ^1.4.1
+  # Flutter Launcher Icons Setup
+  flutter_launcher_icons: ^0.9.0
+  # Flutter Native Splash Screen
+  flutter_native_splash: ^1.1.8+4
+# For information on the generic Dart part of this file, see the
+# following page: https://dart.dev/tools/pub/pubspec
 
-  build_runner: ^1.10.1
-
-  test: ^1.15.5
-
-  inject_generator:
-    git:
-      url: https://github.com/yes-soft-de/inject.dart.git
-      path: package/inject_generator
-
-  # Generate Language Files
-  lang_table: 0.2.0
-  mockito: ^4.1.4
-
-  # Icons Generator
-  flutter_launcher_icons: "^0.7.3"
-
+# The following section is specific to Flutter.
 flutter:
+
+  # The following line ensures that the Material Icons font is
+  # included with your application, so that you can use the icons in
+  # the material Icons class.
   uses-material-design: true
-  assets:
-    - assets/images/
 
-flutter_intl:
-  enabled: true
+  # To add assets to your application, add an assets section, like this:
+  # assets:
+  #   - images/a_dot_burr.jpeg
+  #   - images/a_dot_ham.jpeg
 
-flutter_icons:
-  android: "launcher_icon"
-  ios: true
-  image_path: "assets/images/logo.jpg"
+  # An image asset can refer to one or more resolution-specific "variants", see
+  # https://flutter.dev/assets-and-images/#resolution-aware.
+
+  # For details regarding adding assets from package dependencies, see
+  # https://flutter.dev/assets-and-images/#from-packages
+
+  # To add custom fonts to your application, add a fonts section here,
+  # in this "flutter" section. Each entry in this list should have a
+  # "family" key with the font family name, and a "fonts" key with a
+  # list giving the asset and other descriptors for the font. For
+  # example:
+  # fonts:
+  #   - family: Schyler
+  #     fonts:
+  #       - asset: fonts/Schyler-Regular.ttf
+  #       - asset: fonts/Schyler-Italic.ttf
+  #         style: italic
+  #   - family: Trajan Pro
+  #     fonts:
+  #       - asset: fonts/TrajanPro.ttf
+  #       - asset: fonts/TrajanPro_Bold.ttf
+  #         weight: 700
+  #
+  # For details regarding fonts from package dependencies,
+  # see https://flutter.dev/custom-fonts/#from-packages
+  #  flutter_native_splash:
+  #    color: "#e1f5fe"
+  #    image: assets/images/logo.jpg
+  #    color_dark: "#042a49"
+  #    image_dark: assets/images/logo.jpg
