@@ -1,6 +1,6 @@
 class LoginRequest {
-  String username;
-  String password;
+  String? username;
+  String? password;
 
   LoginRequest({this.username, this.password});
 
@@ -10,7 +10,7 @@ class LoginRequest {
   }
 
   Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = new Map<String, dynamic>();
+    final Map<String, dynamic> data = <String, dynamic>{};
     data['username'] = this.username;
     data['password'] = this.password;
     return data;
