@@ -1,7 +1,7 @@
 class RegisterRequest {
-  String userID;
-  String password;
-  List<String> roles;
+  String? userID;
+  String? password;
+  List<String>? roles;
 
   RegisterRequest({this.userID, this.password, this.roles});
 
@@ -12,7 +12,7 @@ class RegisterRequest {
   }
 
   Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = new Map<String, dynamic>();
+    final Map<String, dynamic> data = <String, dynamic>{};
     data['userID'] = this.userID;
     data['password'] = this.password;
     data['roles'] = this.roles;
