@@ -1,5 +1,5 @@
 class LoginResponse {
-  String token;
+  String? token;
 
   LoginResponse({this.token});
 
@@ -8,7 +8,7 @@ class LoginResponse {
   }
 
   Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = new Map<String, dynamic>();
+    final Map<String, dynamic> data = <String, dynamic>{};
     data['token'] = this.token;
     return data;
   }
