@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
 
 abstract class YesModule {
-  Map<String, WidgetBuilder> getRoutes();
+  static Map<String, WidgetBuilder> RoutesMap = {};
 }
