@@ -1,24 +1,46 @@
 class Urls {
-  static const String DOMAIN = 'http://34.121.141.34';
+  static const String DOMAIN = '';
   static const String BASE_API = DOMAIN + '/html/public';
-  static const String IMAGES_ROOT = DOMAIN + '/upload';
+  static const String IMAGES_ROOT = DOMAIN + '/upload/';
 
-  static const API_SIGN_UP = BASE_API + '/user';
-  static const PROFILE = BASE_API + '/userprofile';
-  static const BRANCH = BASE_API + '/branches';
+  static const SIGN_UP_API = BASE_API + '/user';
+  static const OWNER_PROFILE_API = BASE_API + '/userprofile';
+  static const CAPTAIN_PROFILE_API = BASE_API + '/captainprofile';
+  static const CREATE_BANK_ACCOUNT_API = BASE_API + '/captainprofile';
+  static const BRANCHES_API = BASE_API + '/branches';
   static const CREATE_TOKEN_API = BASE_API + '/login_check';
+  static const LOG_API = BASE_API + '/getRecords';
+  static const DELETE_ORDER = BASE_API + '/ordercancel/';
 
-  static const CREATE_CAPTAIN_ACCOUNT = BASE_API + '/captainprofile';
+  static const REPORT_API = BASE_API + '/report';
 
-  static const NEARBY_ORDERS = BASE_API + '/closestOrders';
-  static const OWNER_ORDERS = BASE_API + '/orders/';
-  static const NEW_ORDER = BASE_API + '/order';
-  static const ORDER_STATUS = BASE_API + '/orderById/';
-  static const ORDER_STATUS_FOR_CAPTAIN =
+  static const NEARBY_ORDERS_API = BASE_API + '/closestOrders';
+  static const OWNER_ORDERS_API = BASE_API + '/orders/';
+  static const CAPTAIN_ACCEPTED_ORDERS_API = BASE_API + '/getAcceptedOrder/';
+  static const NEW_ORDER_API = BASE_API + '/order';
+  static const CAPTAIN_ORDER_UPDATE_API = BASE_API + '/orderUpdateState';
+  static const ACCEPT_ORDER_API = BASE_API + '/acceptedOrder';
+  static const ORDER_STATUS_API = BASE_API + '/orderStatus/';
+  static const ORDER_STATUS_FOR_CAPTAIN_API =
       BASE_API + '/getOrderStatusForCaptain/';
 
-  static const API_UPLOAD = BASE_API + '/upload';
+  static const UPDATE_CAPTAIN_ORDER_API = '/orderUpdateState';
+  static const UPLOAD_API = BASE_API + '/uploadfile';
 
-  static const PACKAGES = BASE_API + '/packages';
-  static const SUBSCRIPTION = BASE_API + '/subscription';
+  static const PACKAGES_API = BASE_API + '/packages';
+  static const SUBSCRIPTION_API = BASE_API + '/subscription';
+  static const RENEW_SUBSCRIPTION_API = BASE_API + '/nextsubscription';
+
+  static const APPOINTMENT_API = BASE_API + '/dating';
+  static const NOTIFICATION_API = BASE_API + '/notificationtoken';
+  static const COMPANYINFO_API = BASE_API + '/companyinfoforuser';
+  static const UPDATES_API = BASE_API + '/updateall';
+  static const ORDER_BY_ID = BASE_API + '/orderStatus/';
+  static const TERMS_CAPTAIN = BASE_API + '/termscaptain';
+  static const TERMS_OWNER = BASE_API + '/';
+  static const NOTIFICATIONNEWCHAT_API = BASE_API + '/notificationnewchat';
+  static const NOTIFICATIONTOADMIN_API = BASE_API + '/notificationtoadmin';
+  static const CAPTAIN_ACTIVE_STATUS_API = BASE_API + '/captainisactive';
+  static const UPDATE_BRANCH_API = BASE_API + '/branches';
+
 }
