@@ -7,7 +7,7 @@
 // ignore_for_file:unnecessary_brace_in_string_interps, unnecessary_new
 // ignore_for_file:prefer_single_quotes,comment_references, directives_ordering
 // ignore_for_file:annotate_overrides,prefer_generic_function_type_aliases
-// ignore_for_file:unused_import, file_names
+// ignore_for_file:unused_import, file_names, avoid_escaping_inner_quotes
 
 import 'package:intl/intl.dart';
 import 'package:intl/message_lookup_by_library.dart';
@@ -20,41 +20,328 @@ class MessageLookup extends MessageLookupByLibrary {
   String get localeName => 'ar';
 
   final messages = _notInlinedMessages(_notInlinedMessages);
-  static _notInlinedMessages(_) => <String, Function> {
-    "acceptOrder" : MessageLookupByLibrary.simpleMessage("Accept Order"),
-    "captain" : MessageLookupByLibrary.simpleMessage("Captain"),
-    "captainGotTheCash" : MessageLookupByLibrary.simpleMessage("Captain got the cash"),
-    "captainIsDelivering" : MessageLookupByLibrary.simpleMessage("Captain is Delivering"),
-    "captainIsInStore" : MessageLookupByLibrary.simpleMessage("Captain is in store"),
-    "captainIsInTheWay" : MessageLookupByLibrary.simpleMessage("Captain is in the way"),
-    "chatRoom" : MessageLookupByLibrary.simpleMessage("Chat Room"),
-    "confirm" : MessageLookupByLibrary.simpleMessage("Confirm"),
-    "confirmCode" : MessageLookupByLibrary.simpleMessage("Confirm Code"),
-    "emailAddressIsRequired" : MessageLookupByLibrary.simpleMessage("Email Address is Required"),
-    "errorHappened" : MessageLookupByLibrary.simpleMessage("Error Happened"),
-    "errorLoadingData" : MessageLookupByLibrary.simpleMessage("Error Loading Data"),
-    "iArrivedAtTheStore" : MessageLookupByLibrary.simpleMessage("I Arrived at the Store"),
-    "iFinishedDelivering" : MessageLookupByLibrary.simpleMessage("I Finished Delivering"),
-    "iGotTheCash" : MessageLookupByLibrary.simpleMessage("I Got the Cash"),
-    "iGotThePackage" : MessageLookupByLibrary.simpleMessage("I Got the Package"),
-    "iHaveAnAccount" : MessageLookupByLibrary.simpleMessage("I have an account"),
-    "language" : MessageLookupByLibrary.simpleMessage("Language"),
-    "lebanon" : MessageLookupByLibrary.simpleMessage("Lebanon"),
-    "loading" : MessageLookupByLibrary.simpleMessage("Loading"),
-    "login" : MessageLookupByLibrary.simpleMessage("Login"),
-    "orderDetails" : MessageLookupByLibrary.simpleMessage("Order Details"),
-    "orderIsDone" : MessageLookupByLibrary.simpleMessage("Order is done!"),
-    "orderIsInUndefinedState" : MessageLookupByLibrary.simpleMessage("Order is in undefined State"),
-    "phoneNumber" : MessageLookupByLibrary.simpleMessage("Phone Number"),
-    "pleaseEnterYourEmail" : MessageLookupByLibrary.simpleMessage("Please Enter Your Email"),
-    "pleaseInputPhoneNumber" : MessageLookupByLibrary.simpleMessage("Please input phone number"),
-    "resendCode" : MessageLookupByLibrary.simpleMessage("Resend Code"),
-    "saudiArabia" : MessageLookupByLibrary.simpleMessage("Saudi Arabia"),
-    "searchingForCaptain" : MessageLookupByLibrary.simpleMessage("Searching for Captain"),
-    "sendMeCode" : MessageLookupByLibrary.simpleMessage("Send me code"),
-    "signOut" : MessageLookupByLibrary.simpleMessage("Sign out"),
-    "startWriting" : MessageLookupByLibrary.simpleMessage("Start Writing"),
-    "storeOwner" : MessageLookupByLibrary.simpleMessage("Store Owner"),
-    "syria" : MessageLookupByLibrary.simpleMessage("Syria")
-  };
+  static Map<String, Function> _notInlinedMessages(_) => <String, Function>{
+        "ProvideDistanceInKm":
+            MessageLookupByLibrary.simpleMessage("المسافة المقطوعة بالكيلومتر"),
+        "accept": MessageLookupByLibrary.simpleMessage("اقبل"),
+        "acceptOrder": MessageLookupByLibrary.simpleMessage("قبول الطلب"),
+        "accountCreated":
+            MessageLookupByLibrary.simpleMessage("تم ارسال الطلب"),
+        "accountNumber": MessageLookupByLibrary.simpleMessage("رقم حساب البنك"),
+        "activeCars": MessageLookupByLibrary.simpleMessage("السيارات المشغولة"),
+        "activePlan": MessageLookupByLibrary.simpleMessage("الباقة الحالية"),
+        "activityLog": MessageLookupByLibrary.simpleMessage("سجل الاستخدام"),
+        "addBranch": MessageLookupByLibrary.simpleMessage("إضافة فرع"),
+        "addBranchFailure": MessageLookupByLibrary.simpleMessage(
+            "عذرا لم تنجح عملية إضافة فرع الرجاء إعادة المحاولة لاحقا"),
+        "addBranchSuccess":
+            MessageLookupByLibrary.simpleMessage("تمت إضافة الفرع بنجاح"),
+        "age": MessageLookupByLibrary.simpleMessage("العمر"),
+        "andIAm": MessageLookupByLibrary.simpleMessage("من أنا"),
+        "apply": MessageLookupByLibrary.simpleMessage("حفظ"),
+        "bankAccountNumber":
+            MessageLookupByLibrary.simpleMessage("رقم الآيبان للحساب البنكي"),
+        "bankName": MessageLookupByLibrary.simpleMessage("اسم البنك"),
+        "bookACar": MessageLookupByLibrary.simpleMessage("احجز سيارة توصيل"),
+        "branch": MessageLookupByLibrary.simpleMessage("الفرع التجاري"),
+        "branch01": MessageLookupByLibrary.simpleMessage("الفرع 01"),
+        "businessName": MessageLookupByLibrary.simpleMessage("الاسم التجاري"),
+        "cancel": MessageLookupByLibrary.simpleMessage("الغاء"),
+        "captain": MessageLookupByLibrary.simpleMessage("كابتن"),
+        "captainAcceptedOrder":
+            MessageLookupByLibrary.simpleMessage("قبل الكابتن الطلب"),
+        "captainGotCash":
+            MessageLookupByLibrary.simpleMessage("حصل الكابتن على الكاش"),
+        "captainGotTheCash":
+            MessageLookupByLibrary.simpleMessage("حصل الكابتن على المال"),
+        "captainInStore":
+            MessageLookupByLibrary.simpleMessage("الكابتن في الفرع التجاري"),
+        "captainIsDelivering":
+            MessageLookupByLibrary.simpleMessage("جار توصيل الطلب"),
+        "captainIsInStore":
+            MessageLookupByLibrary.simpleMessage("الكابتن في المحل"),
+        "captainIsInTheWay":
+            MessageLookupByLibrary.simpleMessage("الكابتن في الطريق"),
+        "captainNotActive": MessageLookupByLibrary.simpleMessage(
+            "لم يتم تفعيل الكابتن من قبل الإدارة بعد الرجاء الانتظار"),
+        "captainStartedDelivery":
+            MessageLookupByLibrary.simpleMessage("بدأ الكابتن التوصيل"),
+        "car": MessageLookupByLibrary.simpleMessage("سيارة "),
+        "cash": MessageLookupByLibrary.simpleMessage("الدفع عند الاستلام"),
+        "changeLanguage": MessageLookupByLibrary.simpleMessage("غير لغة"),
+        "chatRoom": MessageLookupByLibrary.simpleMessage("غرفة المحادثة"),
+        "chatWithStoreOwner":
+            MessageLookupByLibrary.simpleMessage("التحدث مع صاحب المحل"),
+        "checkOrders": MessageLookupByLibrary.simpleMessage("تاكد من الطلبات"),
+        "chooseYourCity":
+            MessageLookupByLibrary.simpleMessage("اختر مدينة تواجدك"),
+        "chooseYourSize":
+            MessageLookupByLibrary.simpleMessage("عدد الموظفين العاملين"),
+        "completedOrders":
+            MessageLookupByLibrary.simpleMessage(" طلبات منتهية "),
+        "confirm": MessageLookupByLibrary.simpleMessage("تأكيد"),
+        "confirmCode": MessageLookupByLibrary.simpleMessage("تأكيد الرقم"),
+        "contactPhoneNumber":
+            MessageLookupByLibrary.simpleMessage("رقم الهاتف للتواصل"),
+        "createNewOrder":
+            MessageLookupByLibrary.simpleMessage("انشاء طلب جديد "),
+        "createNewReport": MessageLookupByLibrary.simpleMessage("دعم سريع"),
+        "created": MessageLookupByLibrary.simpleMessage("تم الانشاء"),
+        "currentBalance":
+            MessageLookupByLibrary.simpleMessage("الرصيد الحالي:"),
+        "currentOrders": MessageLookupByLibrary.simpleMessage("طلباتي"),
+        "darkMode": MessageLookupByLibrary.simpleMessage("الوضع الليلي"),
+        "date": MessageLookupByLibrary.simpleMessage("التاريخ"),
+        "days": MessageLookupByLibrary.simpleMessage("يوم"),
+        "defaultBranch":
+            MessageLookupByLibrary.simpleMessage("الفرع الافتراضي"),
+        "deleteSuccess":
+            MessageLookupByLibrary.simpleMessage("تمت عملية حذف الطلب بنجاح"),
+        "deliver": MessageLookupByLibrary.simpleMessage("وصل طلبات"),
+        "deliverTo": MessageLookupByLibrary.simpleMessage("اسم المستلم"),
+        "directSupport": MessageLookupByLibrary.simpleMessage("دعم مباشر"),
+        "driverLicence": MessageLookupByLibrary.simpleMessage("رخصة القيادة"),
+        "earnCash": MessageLookupByLibrary.simpleMessage("اكسب المال"),
+        "editBranchName":
+            MessageLookupByLibrary.simpleMessage("تعديل اسم الفرع"),
+        "email": MessageLookupByLibrary.simpleMessage("البريد الالكتروني"),
+        "emailAddressIsRequired": MessageLookupByLibrary.simpleMessage(
+            "يرجى توفير البريد الالكتروني"),
+        "errorHappened": MessageLookupByLibrary.simpleMessage("حدث خطأ ما"),
+        "errorLoadingBranches":
+            MessageLookupByLibrary.simpleMessage("حدث خطأ أثناء تحميل الفروع"),
+        "errorLoadingData":
+            MessageLookupByLibrary.simpleMessage("خطأ في تحميل البيانات"),
+        "errorLoggingInFirebaseAccountNotFound":
+            MessageLookupByLibrary.simpleMessage(
+                "Error logging in, firebase account not found"),
+        "errorUploadingImages":
+            MessageLookupByLibrary.simpleMessage("حدث خطأ عند رفع الصور!!"),
+        "finishOrderProvideDistanceInKm": MessageLookupByLibrary.simpleMessage(
+            "انهاء الطلب، المسافة المقطوعة"),
+        "finishedDate": MessageLookupByLibrary.simpleMessage(
+            "لقد انتهت صلاحية اشتراكك الرجاء تجديد الاشتراك من الإعدادات"),
+        "fromWhatsapp": MessageLookupByLibrary.simpleMessage("من واتساب"),
+        "getDirection":
+            MessageLookupByLibrary.simpleMessage("موقع المحل (الاستلام)"),
+        "go": MessageLookupByLibrary.simpleMessage("انطلق"),
+        "help": MessageLookupByLibrary.simpleMessage("Help"),
+        "helpMeMore": MessageLookupByLibrary.simpleMessage("تواصلوا معي"),
+        "home": MessageLookupByLibrary.simpleMessage("الرئيسية"),
+        "hours": MessageLookupByLibrary.simpleMessage("ساعة"),
+        "howWeWork": MessageLookupByLibrary.simpleMessage("كيف يعمل التطبيق"),
+        "iAgreeToTheTermsOfServicePrivacyPolicy":
+            MessageLookupByLibrary.simpleMessage(
+                "اوافق على سياسة الاستخدام والخصوصية"),
+        "iAm": MessageLookupByLibrary.simpleMessage("من أنا"),
+        "iArrivedAtTheStore":
+            MessageLookupByLibrary.simpleMessage("وصلت المحل التجاري"),
+        "iFinishedDelivering":
+            MessageLookupByLibrary.simpleMessage("انتهى الطلب"),
+        "iGotTheCash": MessageLookupByLibrary.simpleMessage("حصلت على المال"),
+        "iGotThePackage":
+            MessageLookupByLibrary.simpleMessage("حصلت على الطلب"),
+        "iHaveAnAccount": MessageLookupByLibrary.simpleMessage("لدي حساب"),
+        "iSpeak": MessageLookupByLibrary.simpleMessage("اللغة المفضلة"),
+        "inactive": MessageLookupByLibrary.simpleMessage(
+            "اشتراكك غير مفعل بعد الرجاء انتظار التفعيل من قبل الإدارة"),
+        "info": MessageLookupByLibrary.simpleMessage("حول الطلب"),
+        "initData": MessageLookupByLibrary.simpleMessage("تعيين البيانات"),
+        "language": MessageLookupByLibrary.simpleMessage("اللغة"),
+        "largeMoreThan100Employees":
+            MessageLookupByLibrary.simpleMessage("+100 موظف"),
+        "latestUpdates": MessageLookupByLibrary.simpleMessage("اخر التحديثات"),
+        "lebanon": MessageLookupByLibrary.simpleMessage("لبنان"),
+        "loading": MessageLookupByLibrary.simpleMessage("جار التحميل"),
+        "locationOfCustomer":
+            MessageLookupByLibrary.simpleMessage("(التسليم) موقع الزبون"),
+        "login": MessageLookupByLibrary.simpleMessage("تسجيل الدخول"),
+        "me": MessageLookupByLibrary.simpleMessage("مهنتي"),
+        "mediumMoreThan20EmployeesLessThan100":
+            MessageLookupByLibrary.simpleMessage("21 - 100 موظف"),
+        "minutes": MessageLookupByLibrary.simpleMessage("دقيقة"),
+        "mohammad": MessageLookupByLibrary.simpleMessage("محمد"),
+        "moveToOrders":
+            MessageLookupByLibrary.simpleMessage("الانتقال للطلبات"),
+        "myBalance": MessageLookupByLibrary.simpleMessage("رصيدي"),
+        "myOrders": MessageLookupByLibrary.simpleMessage("طلباتي"),
+        "myPlan": MessageLookupByLibrary.simpleMessage("باقتي"),
+        "myProfile": MessageLookupByLibrary.simpleMessage("معلوماتي"),
+        "myStatus": MessageLookupByLibrary.simpleMessage("حالتي"),
+        "name": MessageLookupByLibrary.simpleMessage("الاسم"),
+        "nameIsRequired":
+            MessageLookupByLibrary.simpleMessage("حقل الاسم فارغ"),
+        "nearbyOrders": MessageLookupByLibrary.simpleMessage("الطلبات القريبة"),
+        "newName": MessageLookupByLibrary.simpleMessage("اسم جديد"),
+        "newOrder": MessageLookupByLibrary.simpleMessage("طلب جديد"),
+        "next": MessageLookupByLibrary.simpleMessage("التالي"),
+        "nextPaymentDate":
+            MessageLookupByLibrary.simpleMessage("تاريخ الدفعة القادمة:"),
+        "noImage": MessageLookupByLibrary.simpleMessage("لا يوجد صورة"),
+        "notSubscription": MessageLookupByLibrary.simpleMessage(
+            "لم تشترك بباقة بعد الرجاء الاشتراك من الإعدادات"),
+        "notVerified":
+            MessageLookupByLibrary.simpleMessage("لم يتم تفعيل الباقة بعد"),
+        "notifications": MessageLookupByLibrary.simpleMessage("الاشعارات"),
+        "online": MessageLookupByLibrary.simpleMessage("مدفوع"),
+        "openChatRoom":
+            MessageLookupByLibrary.simpleMessage("افتح غرفة المحادثة"),
+        "openTheApp": MessageLookupByLibrary.simpleMessage("افتح التطبيق"),
+        "order": MessageLookupByLibrary.simpleMessage("الطلب "),
+        "orderCreatedReturnToOrders":
+            MessageLookupByLibrary.simpleMessage("انتهى انشاء الطلب"),
+        "orderDetails": MessageLookupByLibrary.simpleMessage("تفاصيل الطلب"),
+        "orderIsCreated": MessageLookupByLibrary.simpleMessage("تم الانشاء"),
+        "orderIsDone": MessageLookupByLibrary.simpleMessage("انتهى الطلب!"),
+        "orderIsFinished":
+            MessageLookupByLibrary.simpleMessage("تم انهاء الطلب"),
+        "orderIsInUndefinedState":
+            MessageLookupByLibrary.simpleMessage("الطلب في مرحلة غير معروفة"),
+        "ordermonth": MessageLookupByLibrary.simpleMessage("طلب في الشهر"),
+        "orders": MessageLookupByLibrary.simpleMessage("الطلبات"),
+        "ordersMonth": MessageLookupByLibrary.simpleMessage("طلب / الشهر"),
+        "ourBankAccountNumber":
+            MessageLookupByLibrary.simpleMessage("رقم حسابنا"),
+        "ourBankName": MessageLookupByLibrary.simpleMessage("اسم بنكنا"),
+        "ourPackages": MessageLookupByLibrary.simpleMessage("باقاتنا"),
+        "outOfCars": MessageLookupByLibrary.simpleMessage(
+            "لا يوجد سيارات متوفرة لك الرجاء انتظار طلباتك الجارية من الانتهاء"),
+        "outOforders": MessageLookupByLibrary.simpleMessage(
+            "لقد استهلكت الحد الأقصى من الطلبات يمكنك تجديد الاشتراك من الإعدادات"),
+        "package": MessageLookupByLibrary.simpleMessage("باقة"),
+        "password": MessageLookupByLibrary.simpleMessage("كلمة السر"),
+        "passwordIsTooShort":
+            MessageLookupByLibrary.simpleMessage("كلمة السر قصيرة جدا"),
+        "paySubscription": MessageLookupByLibrary.simpleMessage("دفع الاشتراك"),
+        "paymentHistory": MessageLookupByLibrary.simpleMessage("تاريخ الدفعات"),
+        "paymentMethod": MessageLookupByLibrary.simpleMessage("طريقة الدفع"),
+        "phoneIsRequired":
+            MessageLookupByLibrary.simpleMessage("الرجاء تزويدنا برقم الهاتف"),
+        "phoneNumber": MessageLookupByLibrary.simpleMessage("رقم الهاتف"),
+        "phoneNumbertooShort":
+            MessageLookupByLibrary.simpleMessage("رقم الهاتف المدخل قصير"),
+        "pickAJob": MessageLookupByLibrary.simpleMessage("اختار عملي"),
+        "pickALanguage": MessageLookupByLibrary.simpleMessage("اختار لغتي"),
+        "pleaseCompleteTheForm":
+            MessageLookupByLibrary.simpleMessage("الرجاء أكمال الفورم"),
+        "pleaseDownloadpasco_shipping":
+            MessageLookupByLibrary.simpleMessage(
+                "الرجاء تحميل pasco_shipping"),
+        "pleaseEnterYourEmail": MessageLookupByLibrary.simpleMessage(
+            "الرجاء ادخال البريد الالكتروني"),
+        "pleaseInputPhoneNumber":
+            MessageLookupByLibrary.simpleMessage("الرجاء ادخال رقم الهاتف"),
+        "pleaseProvideTheDistance": MessageLookupByLibrary.simpleMessage(
+            "الرجاء توفير المسافة المقطوعة"),
+        "pleaseProvideUsTheClientPhoneNumber":
+            MessageLookupByLibrary.simpleMessage(
+                "الرجاء تزويدنا رقم الزبون او التخطي"),
+        "pleaseProvideUsWithTheClientName":
+            MessageLookupByLibrary.simpleMessage(
+                "الرجاء تزويدنا باسم الزبون، او التخطي"),
+        "pleaseSelectABranch":
+            MessageLookupByLibrary.simpleMessage("الرجاء اختيار فرع تجاري"),
+        "privacyPolicy": MessageLookupByLibrary.simpleMessage("شروط الخصوصية"),
+        "profile": MessageLookupByLibrary.simpleMessage("الملف الشخصي"),
+        "reasonIsRequired":
+            MessageLookupByLibrary.simpleMessage("الرجاء تزويدنا بسبب الابلاغ"),
+        "reasonOfTheReport":
+            MessageLookupByLibrary.simpleMessage("سبب طلب الدعم"),
+        "recipientName": MessageLookupByLibrary.simpleMessage("اسم المستلم"),
+        "recipientPhoneNumber":
+            MessageLookupByLibrary.simpleMessage("رقم هاتف المستلم"),
+        "register": MessageLookupByLibrary.simpleMessage("انشاء حساب جديد"),
+        "registerSuccessSetupMyProfile": MessageLookupByLibrary.simpleMessage(
+            "تم تأكيد الحساب، اتجه للاشتراك"),
+        "renewNewPlan":
+            MessageLookupByLibrary.simpleMessage("اشتراك بباقة جديدة"),
+        "renewOldPlan":
+            MessageLookupByLibrary.simpleMessage("تجديد الاشتراك الحالي"),
+        "renewPlan":
+            MessageLookupByLibrary.simpleMessage("كيفية تجديد الإشتراك"),
+        "renewSubscription":
+            MessageLookupByLibrary.simpleMessage("تجديد الاشتراك"),
+        "reportSent": MessageLookupByLibrary.simpleMessage("تم ارسال الابلاغ"),
+        "requestMeeting":
+            MessageLookupByLibrary.simpleMessage("اطلب موعدا للقاء"),
+        "resendCode": MessageLookupByLibrary.simpleMessage("اعادة الارسال"),
+        "saudiArabia": MessageLookupByLibrary.simpleMessage("السعودية"),
+        "saudiRiyal": MessageLookupByLibrary.simpleMessage("ريال سعودي"),
+        "save": MessageLookupByLibrary.simpleMessage("الحفظ"),
+        "saveBranch": MessageLookupByLibrary.simpleMessage("حفظ الفرع"),
+        "saveBranches": MessageLookupByLibrary.simpleMessage("حفظ الفروع"),
+        "saveLocationAsBranch01":
+            MessageLookupByLibrary.simpleMessage("احفظ مكان الفرع 01"),
+        "saveSuccess": MessageLookupByLibrary.simpleMessage("تم بنجاح"),
+        "searchingForCaptain":
+            MessageLookupByLibrary.simpleMessage("جار البحث عن كابتن"),
+        "seconds": MessageLookupByLibrary.simpleMessage("ثانية"),
+        "sendMeCode": MessageLookupByLibrary.simpleMessage("ارسل رمز التأكيد"),
+        "settings": MessageLookupByLibrary.simpleMessage("الاعدادات"),
+        "share": MessageLookupByLibrary.simpleMessage("المشاركة"),
+        "signOut": MessageLookupByLibrary.simpleMessage("تسحيل الخروج"),
+        "skip": MessageLookupByLibrary.simpleMessage("تخطي"),
+        "smallLessThan20Employee":
+            MessageLookupByLibrary.simpleMessage("1 - 20 موظف"),
+        "startWriting": MessageLookupByLibrary.simpleMessage("ابدأ الكتابة"),
+        "stcPayCode": MessageLookupByLibrary.simpleMessage("رقم حساب STC PAY"),
+        "store": MessageLookupByLibrary.simpleMessage("المتجر"),
+        "storeName": MessageLookupByLibrary.simpleMessage("اسم المحل التجاري"),
+        "storeOwner": MessageLookupByLibrary.simpleMessage("محل تجاري"),
+        "storePhone": MessageLookupByLibrary.simpleMessage("رقم هاتف المحل"),
+        "submittingProfile":
+            MessageLookupByLibrary.simpleMessage("جار حفظ الملف"),
+        "subscribe": MessageLookupByLibrary.simpleMessage("اشترك"),
+        "successRenew":
+            MessageLookupByLibrary.simpleMessage("تم تجديد الباقة بنجاح"),
+        "sureForDelete": MessageLookupByLibrary.simpleMessage(
+            "هل أنت متأكد من قيامك بحذف هذا الطلب"),
+        "syria": MessageLookupByLibrary.simpleMessage("سوريا"),
+        "termsOfService":
+            MessageLookupByLibrary.simpleMessage("شروط الاستخدام"),
+        "thisMightTakeAWhilePleaseWait": MessageLookupByLibrary.simpleMessage(
+            "سوف يستغرق الامر دقيقة واحدة الرجاء الانتظار"),
+        "to": MessageLookupByLibrary.simpleMessage("الى"),
+        "toFindOutMorePleaseLeaveYourPhonenandWeWill":
+            MessageLookupByLibrary.simpleMessage(
+                "للمزيد من المعلومات خذ موعدا بترك رقم هاتفك، وسيتم التواصل معك"),
+        "took": MessageLookupByLibrary.simpleMessage("استغرق"),
+        "unaccept": MessageLookupByLibrary.simpleMessage(
+            "لم يتم الموافقة على اشتراكك من قبل الإدارة يمكنك التواصل معنا للحصول على المزيد من التفاصيل"),
+        "unknown": MessageLookupByLibrary.simpleMessage("غير معروف"),
+        "unknownNumberOfCar":
+            MessageLookupByLibrary.simpleMessage("عدد السيارات غير محدود"),
+        "updateBranch": MessageLookupByLibrary.simpleMessage("تعديل الفرع"),
+        "updateBranchFailure": MessageLookupByLibrary.simpleMessage(
+            "عذرا لم يتم تعديل الفرع بنجاح الرجاء إعادة المحاولة لاحقا"),
+        "updateBranchSuccess":
+            MessageLookupByLibrary.simpleMessage("تم تعديل الفرع بنجاح"),
+        "updateBranches": MessageLookupByLibrary.simpleMessage("تعديل الأفرع"),
+        "uploadAndSubmit":
+            MessageLookupByLibrary.simpleMessage("رفع الصور وارسال الطلب"),
+        "uploadingAndSubmitting":
+            MessageLookupByLibrary.simpleMessage("جار الرفع وارسال الطلب"),
+        "uploadingImages":
+            MessageLookupByLibrary.simpleMessage("جار رفع الصور"),
+        "uploadingImagesPleaseWait":
+            MessageLookupByLibrary.simpleMessage("يتم رفع الصور"),
+        "warnning": MessageLookupByLibrary.simpleMessage("ملاحظة هامة"),
+        "weDeliver":
+            MessageLookupByLibrary.simpleMessage("نوصل الطلب بدلا عنك"),
+        "weNeedYourImageAndYourDriverLicenceInOrder":
+            MessageLookupByLibrary.simpleMessage(
+                "نحتاج صورتك الشخصية وصورة رخصة القيادة للتسجيل"),
+        "weNeedYourImageInOrderToSignUp":
+            MessageLookupByLibrary.simpleMessage("نحتاج صورتك للتسجيل"),
+        "weWillContactYouSoon": MessageLookupByLibrary.simpleMessage(
+            "تم تسجيل الطلب، سنتواصل معك بأقرب وقت ممكن"),
+        "welcomeToOrdersScreen":
+            MessageLookupByLibrary.simpleMessage("اهلا في قسم الطلبات"),
+        "welcomeTopasco_shipping": MessageLookupByLibrary.simpleMessage(
+            "اهلا بك في pasco_shipping"),
+        "whatsappWithCaptain": MessageLookupByLibrary.simpleMessage("الكابتن"),
+        "whatsappWithClient": MessageLookupByLibrary.simpleMessage("الزبون"),
+        "whatsappWithStoreOwner":
+            MessageLookupByLibrary.simpleMessage("صاحب المحل")
+      };
 }
