@@ -7,7 +7,7 @@
 // ignore_for_file:unnecessary_brace_in_string_interps, unnecessary_new
 // ignore_for_file:prefer_single_quotes,comment_references, directives_ordering
 // ignore_for_file:annotate_overrides,prefer_generic_function_type_aliases
-// ignore_for_file:unused_import, file_names
+// ignore_for_file:unused_import, file_names, avoid_escaping_inner_quotes
 
 import 'package:intl/intl.dart';
 import 'package:intl/message_lookup_by_library.dart';
@@ -20,41 +20,335 @@ class MessageLookup extends MessageLookupByLibrary {
   String get localeName => 'en';
 
   final messages = _notInlinedMessages(_notInlinedMessages);
-  static _notInlinedMessages(_) => <String, Function> {
-    "acceptOrder" : MessageLookupByLibrary.simpleMessage("Accept Order"),
-    "captain" : MessageLookupByLibrary.simpleMessage("Captain"),
-    "captainGotTheCash" : MessageLookupByLibrary.simpleMessage("Captain got the cash"),
-    "captainIsDelivering" : MessageLookupByLibrary.simpleMessage("Captain is Delivering"),
-    "captainIsInStore" : MessageLookupByLibrary.simpleMessage("Captain is in store"),
-    "captainIsInTheWay" : MessageLookupByLibrary.simpleMessage("Captain is in the way"),
-    "chatRoom" : MessageLookupByLibrary.simpleMessage("Chat Room"),
-    "confirm" : MessageLookupByLibrary.simpleMessage("Confirm"),
-    "confirmCode" : MessageLookupByLibrary.simpleMessage("Confirm Code"),
-    "emailAddressIsRequired" : MessageLookupByLibrary.simpleMessage("Email Address is Required"),
-    "errorHappened" : MessageLookupByLibrary.simpleMessage("Error Happened"),
-    "errorLoadingData" : MessageLookupByLibrary.simpleMessage("Error Loading Data"),
-    "iArrivedAtTheStore" : MessageLookupByLibrary.simpleMessage("I Arrived at the Store"),
-    "iFinishedDelivering" : MessageLookupByLibrary.simpleMessage("I Finished Delivering"),
-    "iGotTheCash" : MessageLookupByLibrary.simpleMessage("I Got the Cash"),
-    "iGotThePackage" : MessageLookupByLibrary.simpleMessage("I Got the Package"),
-    "iHaveAnAccount" : MessageLookupByLibrary.simpleMessage("I have an account"),
-    "language" : MessageLookupByLibrary.simpleMessage("Language"),
-    "lebanon" : MessageLookupByLibrary.simpleMessage("Lebanon"),
-    "loading" : MessageLookupByLibrary.simpleMessage("Loading"),
-    "login" : MessageLookupByLibrary.simpleMessage("Login"),
-    "orderDetails" : MessageLookupByLibrary.simpleMessage("Order Details"),
-    "orderIsDone" : MessageLookupByLibrary.simpleMessage("Order is done!"),
-    "orderIsInUndefinedState" : MessageLookupByLibrary.simpleMessage("Order is in undefined State"),
-    "phoneNumber" : MessageLookupByLibrary.simpleMessage("Phone Number"),
-    "pleaseEnterYourEmail" : MessageLookupByLibrary.simpleMessage("Please Enter Your Email"),
-    "pleaseInputPhoneNumber" : MessageLookupByLibrary.simpleMessage("Please input phone number"),
-    "resendCode" : MessageLookupByLibrary.simpleMessage("Resend Code"),
-    "saudiArabia" : MessageLookupByLibrary.simpleMessage("Saudi Arabia"),
-    "searchingForCaptain" : MessageLookupByLibrary.simpleMessage("Searching for Captain"),
-    "sendMeCode" : MessageLookupByLibrary.simpleMessage("Send me code"),
-    "signOut" : MessageLookupByLibrary.simpleMessage("Sign out"),
-    "startWriting" : MessageLookupByLibrary.simpleMessage("Start Writing"),
-    "storeOwner" : MessageLookupByLibrary.simpleMessage("Store Owner"),
-    "syria" : MessageLookupByLibrary.simpleMessage("Syria")
-  };
+  static Map<String, Function> _notInlinedMessages(_) => <String, Function>{
+        "ProvideDistanceInKm":
+            MessageLookupByLibrary.simpleMessage("distance in KM"),
+        "accept": MessageLookupByLibrary.simpleMessage("Accept"),
+        "acceptOrder": MessageLookupByLibrary.simpleMessage("Accept Order"),
+        "accountCreated":
+            MessageLookupByLibrary.simpleMessage("Account Created"),
+        "accountNumber": MessageLookupByLibrary.simpleMessage("Account Number"),
+        "activeCars": MessageLookupByLibrary.simpleMessage("Active Cars"),
+        "activePlan": MessageLookupByLibrary.simpleMessage("Active Plan"),
+        "activityLog": MessageLookupByLibrary.simpleMessage("Activity Log"),
+        "addBranch": MessageLookupByLibrary.simpleMessage("Add Branch"),
+        "addBranchFailure": MessageLookupByLibrary.simpleMessage(
+            "Sorry your request for add branch faild please try again"),
+        "addBranchSuccess": MessageLookupByLibrary.simpleMessage(
+            "Branch has been added successfully"),
+        "age": MessageLookupByLibrary.simpleMessage("Age"),
+        "andIAm": MessageLookupByLibrary.simpleMessage("And I am"),
+        "apply": MessageLookupByLibrary.simpleMessage("APPLY"),
+        "bankAccountNumber":
+            MessageLookupByLibrary.simpleMessage("Bank Account Number"),
+        "bankName": MessageLookupByLibrary.simpleMessage("Bank Name"),
+        "bookACar": MessageLookupByLibrary.simpleMessage("Book a car"),
+        "branch": MessageLookupByLibrary.simpleMessage("Branch"),
+        "branch01": MessageLookupByLibrary.simpleMessage("Branch 01"),
+        "businessName": MessageLookupByLibrary.simpleMessage("Business Name"),
+        "cancel": MessageLookupByLibrary.simpleMessage("CANCEL"),
+        "captain": MessageLookupByLibrary.simpleMessage("Captain"),
+        "captainAcceptedOrder":
+            MessageLookupByLibrary.simpleMessage("Captain Accepted Order"),
+        "captainGotCash":
+            MessageLookupByLibrary.simpleMessage("Captain Got Cash"),
+        "captainGotTheCash":
+            MessageLookupByLibrary.simpleMessage("Captain got the cash"),
+        "captainInStore":
+            MessageLookupByLibrary.simpleMessage("Captain in Store"),
+        "captainIsDelivering":
+            MessageLookupByLibrary.simpleMessage("Captain is Delivering"),
+        "captainIsInStore":
+            MessageLookupByLibrary.simpleMessage("Captain is in store"),
+        "captainIsInTheWay":
+            MessageLookupByLibrary.simpleMessage("Captain is in the way"),
+        "captainNotActive": MessageLookupByLibrary.simpleMessage(
+            "Your account not activated yet please wait"),
+        "captainStartedDelivery":
+            MessageLookupByLibrary.simpleMessage("Captain Started Delivery"),
+        "car": MessageLookupByLibrary.simpleMessage("Car"),
+        "cash": MessageLookupByLibrary.simpleMessage("Cash"),
+        "changeLanguage":
+            MessageLookupByLibrary.simpleMessage("Change Language"),
+        "chatRoom": MessageLookupByLibrary.simpleMessage("Chat Room"),
+        "chatWithStoreOwner":
+            MessageLookupByLibrary.simpleMessage("Chat with Store Owner"),
+        "checkOrders": MessageLookupByLibrary.simpleMessage("Check Orders"),
+        "chooseYourCity":
+            MessageLookupByLibrary.simpleMessage("Choose Your City"),
+        "chooseYourSize":
+            MessageLookupByLibrary.simpleMessage("Choose Your Size"),
+        "completedOrders":
+            MessageLookupByLibrary.simpleMessage("Completed Orders"),
+        "confirm": MessageLookupByLibrary.simpleMessage("Confirm"),
+        "confirmCode": MessageLookupByLibrary.simpleMessage("Confirm Code"),
+        "contactPhoneNumber":
+            MessageLookupByLibrary.simpleMessage("Contact Phone Number"),
+        "createNewOrder":
+            MessageLookupByLibrary.simpleMessage("Create new order"),
+        "createNewReport":
+            MessageLookupByLibrary.simpleMessage("Create new Report"),
+        "created": MessageLookupByLibrary.simpleMessage("Created"),
+        "currentBalance":
+            MessageLookupByLibrary.simpleMessage("Current Balance:"),
+        "currentOrders": MessageLookupByLibrary.simpleMessage("Current Orders"),
+        "darkMode": MessageLookupByLibrary.simpleMessage("Dark Mode"),
+        "date": MessageLookupByLibrary.simpleMessage("Date"),
+        "days": MessageLookupByLibrary.simpleMessage("day"),
+        "defaultBranch": MessageLookupByLibrary.simpleMessage("Default Branch"),
+        "deleteSuccess": MessageLookupByLibrary.simpleMessage(
+            "the order has been deleted successfully"),
+        "deliver": MessageLookupByLibrary.simpleMessage("Deliver"),
+        "deliverTo": MessageLookupByLibrary.simpleMessage("Deliver To"),
+        "directSupport": MessageLookupByLibrary.simpleMessage("Direct Support"),
+        "driverLicence": MessageLookupByLibrary.simpleMessage("Driver Licence"),
+        "earnCash": MessageLookupByLibrary.simpleMessage("Earn Cash"),
+        "editBranchName":
+            MessageLookupByLibrary.simpleMessage("Edit Branch Name"),
+        "email": MessageLookupByLibrary.simpleMessage("Email"),
+        "emailAddressIsRequired":
+            MessageLookupByLibrary.simpleMessage("Email Address is Required"),
+        "errorHappened": MessageLookupByLibrary.simpleMessage("Error Happened"),
+        "errorLoadingBranches":
+            MessageLookupByLibrary.simpleMessage("Error Loading Branches"),
+        "errorLoadingData":
+            MessageLookupByLibrary.simpleMessage("Error Loading Data"),
+        "errorLoggingInFirebaseAccountNotFound":
+            MessageLookupByLibrary.simpleMessage(
+                "Error logging in, firebase account not found"),
+        "errorUploadingImages":
+            MessageLookupByLibrary.simpleMessage("Error Uploading Images!!"),
+        "failureRenew": MessageLookupByLibrary.simpleMessage(
+            "Your renew has been failed please try again"),
+        "finishOrderProvideDistanceInKm": MessageLookupByLibrary.simpleMessage(
+            "Finish Order, Provide distance in KM"),
+        "finishedDate": MessageLookupByLibrary.simpleMessage(
+            "Your subscription is out of date please renew your subscription"),
+        "fromWhatsapp": MessageLookupByLibrary.simpleMessage("From WhatsApp"),
+        "getDirection": MessageLookupByLibrary.simpleMessage("Get Direction"),
+        "go": MessageLookupByLibrary.simpleMessage("Go"),
+        "help": MessageLookupByLibrary.simpleMessage("Help"),
+        "helpMeMore": MessageLookupByLibrary.simpleMessage("Help me more"),
+        "home": MessageLookupByLibrary.simpleMessage("Home"),
+        "hours": MessageLookupByLibrary.simpleMessage("hour"),
+        "howWeWork": MessageLookupByLibrary.simpleMessage("How we work"),
+        "iAgreeToTheTermsOfServicePrivacyPolicy":
+            MessageLookupByLibrary.simpleMessage(
+                "I Agree To The Terms Of Service & privacy policy"),
+        "iAgreeToTheTermsOsService": MessageLookupByLibrary.simpleMessage(
+            "I Agree to the Terms os service."),
+        "iAm": MessageLookupByLibrary.simpleMessage("I am"),
+        "iArrivedAtTheStore":
+            MessageLookupByLibrary.simpleMessage("I Arrived at the Store"),
+        "iFinishedDelivering":
+            MessageLookupByLibrary.simpleMessage("I Finished Delivering"),
+        "iGotTheCash": MessageLookupByLibrary.simpleMessage("I Got the Cash"),
+        "iGotThePackage":
+            MessageLookupByLibrary.simpleMessage("I Got the Package"),
+        "iHaveAnAccount":
+            MessageLookupByLibrary.simpleMessage("I have an account"),
+        "iSpeak": MessageLookupByLibrary.simpleMessage("Preferred language"),
+        "inactive": MessageLookupByLibrary.simpleMessage(
+            "Your subscription not activate yet please wait"),
+        "info": MessageLookupByLibrary.simpleMessage("Info"),
+        "initData": MessageLookupByLibrary.simpleMessage("Init Data"),
+        "language": MessageLookupByLibrary.simpleMessage("Language"),
+        "largeMoreThan100Employees":
+            MessageLookupByLibrary.simpleMessage("+100 Employees"),
+        "latestUpdates": MessageLookupByLibrary.simpleMessage("Latest Updates"),
+        "lebanon": MessageLookupByLibrary.simpleMessage("Lebanon"),
+        "loading": MessageLookupByLibrary.simpleMessage("Loading"),
+        "locationOfCustomer":
+            MessageLookupByLibrary.simpleMessage("Location of customer"),
+        "login": MessageLookupByLibrary.simpleMessage("Login"),
+        "me": MessageLookupByLibrary.simpleMessage("ME"),
+        "mediumMoreThan20EmployeesLessThan100":
+            MessageLookupByLibrary.simpleMessage("21 - 100 Employees"),
+        "minutes": MessageLookupByLibrary.simpleMessage("minute"),
+        "mohammad": MessageLookupByLibrary.simpleMessage("Mohammad"),
+        "moveToOrders": MessageLookupByLibrary.simpleMessage("Move to Orders"),
+        "myBalance": MessageLookupByLibrary.simpleMessage("My Balance"),
+        "myOrders": MessageLookupByLibrary.simpleMessage("My Orders"),
+        "myPlan": MessageLookupByLibrary.simpleMessage("My Plan"),
+        "myProfile": MessageLookupByLibrary.simpleMessage("My Profile"),
+        "myStatus": MessageLookupByLibrary.simpleMessage("My Status"),
+        "name": MessageLookupByLibrary.simpleMessage("Name"),
+        "nameIsRequired":
+            MessageLookupByLibrary.simpleMessage("Name is required"),
+        "nearbyOrders": MessageLookupByLibrary.simpleMessage("Nearby Orders"),
+        "newName": MessageLookupByLibrary.simpleMessage("New name"),
+        "newOrder": MessageLookupByLibrary.simpleMessage("New Order"),
+        "next": MessageLookupByLibrary.simpleMessage("Next"),
+        "nextPaymentDate":
+            MessageLookupByLibrary.simpleMessage("Next Payment Date:"),
+        "noImage": MessageLookupByLibrary.simpleMessage("No Image"),
+        "notSubscription": MessageLookupByLibrary.simpleMessage(
+            "You dont have a subscription yet"),
+        "notVerified": MessageLookupByLibrary.simpleMessage(
+            "Your subscription not verified yet"),
+        "notifications": MessageLookupByLibrary.simpleMessage("Notifications"),
+        "online": MessageLookupByLibrary.simpleMessage("online"),
+        "openChatRoom": MessageLookupByLibrary.simpleMessage("Open Chat Room"),
+        "openTheApp": MessageLookupByLibrary.simpleMessage("Open the app"),
+        "order": MessageLookupByLibrary.simpleMessage("Order"),
+        "orderCreatedReturnToOrders": MessageLookupByLibrary.simpleMessage(
+            "Order Created, Return to Orders"),
+        "orderDetails": MessageLookupByLibrary.simpleMessage("Order Details"),
+        "orderIsCreated":
+            MessageLookupByLibrary.simpleMessage("Order is created"),
+        "orderIsDone": MessageLookupByLibrary.simpleMessage("Order is done!"),
+        "orderIsFinished":
+            MessageLookupByLibrary.simpleMessage("Order is finished"),
+        "orderIsInUndefinedState":
+            MessageLookupByLibrary.simpleMessage("Order is in undefined State"),
+        "ordermonth": MessageLookupByLibrary.simpleMessage("Order/Month"),
+        "orders": MessageLookupByLibrary.simpleMessage("Orders"),
+        "ordersMonth": MessageLookupByLibrary.simpleMessage("Orders / Month"),
+        "ourBankAccountNumber":
+            MessageLookupByLibrary.simpleMessage("Our Bank Account Number"),
+        "ourBankName": MessageLookupByLibrary.simpleMessage("Our Bank Name"),
+        "ourPackages": MessageLookupByLibrary.simpleMessage("Our Packages"),
+        "outOfCars": MessageLookupByLibrary.simpleMessage(
+            "There is no car available please wait your ongoing orders to finish"),
+        "outOforders": MessageLookupByLibrary.simpleMessage(
+            "You reach your limit of orders you can renew your subscription from settings"),
+        "package": MessageLookupByLibrary.simpleMessage("Package"),
+        "password": MessageLookupByLibrary.simpleMessage("Password"),
+        "passwordIsTooShort":
+            MessageLookupByLibrary.simpleMessage("Password is too short"),
+        "paySubscription":
+            MessageLookupByLibrary.simpleMessage("Pay Subscription"),
+        "paymentHistory":
+            MessageLookupByLibrary.simpleMessage("Payment History"),
+        "paymentMethod": MessageLookupByLibrary.simpleMessage("Payment Method"),
+        "phoneIsRequired":
+            MessageLookupByLibrary.simpleMessage("Phone is required"),
+        "phoneNumber": MessageLookupByLibrary.simpleMessage("Phone Number"),
+        "phoneNumbertooShort":
+            MessageLookupByLibrary.simpleMessage("Phone number is too short"),
+        "pickAJob": MessageLookupByLibrary.simpleMessage("Pick a Job"),
+        "pickALanguage":
+            MessageLookupByLibrary.simpleMessage("Pick a Language"),
+        "pleaseCompleteTheForm":
+            MessageLookupByLibrary.simpleMessage("Please Complete the Form"),
+        "pleaseDownloadpasco_shipping":
+            MessageLookupByLibrary.simpleMessage(
+                "Please Download pasco_shipping"),
+        "pleaseEnterYourEmail":
+            MessageLookupByLibrary.simpleMessage("Please Enter Your Email"),
+        "pleaseInputPhoneNumber":
+            MessageLookupByLibrary.simpleMessage("Please input phone number"),
+        "pleaseProvideTheDistance":
+            MessageLookupByLibrary.simpleMessage("Please Provide the distance"),
+        "pleaseProvideUsTheClientPhoneNumber":
+            MessageLookupByLibrary.simpleMessage(
+                "Please Provide us the client phone number"),
+        "pleaseProvideUsWithTheClientName":
+            MessageLookupByLibrary.simpleMessage(
+                "Please Provide us with the Client Name"),
+        "pleaseSelectABranch":
+            MessageLookupByLibrary.simpleMessage("Please Select a Branch"),
+        "privacyPolicy": MessageLookupByLibrary.simpleMessage("Privacy Policy"),
+        "profile": MessageLookupByLibrary.simpleMessage("Profile"),
+        "reasonIsRequired":
+            MessageLookupByLibrary.simpleMessage("Reason is Required"),
+        "reasonOfTheReport":
+            MessageLookupByLibrary.simpleMessage("Reason of the report"),
+        "recipientName": MessageLookupByLibrary.simpleMessage("Recipient Name"),
+        "recipientPhoneNumber":
+            MessageLookupByLibrary.simpleMessage("Recipient Phone Number"),
+        "register": MessageLookupByLibrary.simpleMessage("Register"),
+        "registerSuccessSetupMyProfile": MessageLookupByLibrary.simpleMessage(
+            "Register Success, Setup my profile"),
+        "renewNewPlan": MessageLookupByLibrary.simpleMessage(
+            "Subscription with new package"),
+        "renewOldPlan":
+            MessageLookupByLibrary.simpleMessage("Renew old package"),
+        "renewPlan":
+            MessageLookupByLibrary.simpleMessage("Subscription options"),
+        "renewSubscription":
+            MessageLookupByLibrary.simpleMessage("Renew Subscription"),
+        "reportSent": MessageLookupByLibrary.simpleMessage("Report sent"),
+        "requestMeeting":
+            MessageLookupByLibrary.simpleMessage("Request Meeting"),
+        "resendCode": MessageLookupByLibrary.simpleMessage("Resend Code"),
+        "saudiArabia": MessageLookupByLibrary.simpleMessage("Saudi Arabia"),
+        "saudiRiyal": MessageLookupByLibrary.simpleMessage("Saudi Riyal"),
+        "save": MessageLookupByLibrary.simpleMessage("Save"),
+        "saveBranch": MessageLookupByLibrary.simpleMessage("Save Branch"),
+        "saveBranches": MessageLookupByLibrary.simpleMessage("Save Branches"),
+        "saveLocationAsBranch01":
+            MessageLookupByLibrary.simpleMessage("Save Location as Branch 01"),
+        "saveSuccess": MessageLookupByLibrary.simpleMessage("Save Success"),
+        "searchingForCaptain":
+            MessageLookupByLibrary.simpleMessage("Searching for Captain"),
+        "seconds": MessageLookupByLibrary.simpleMessage("second"),
+        "sendMeCode": MessageLookupByLibrary.simpleMessage("Send me code"),
+        "settings": MessageLookupByLibrary.simpleMessage("Settings"),
+        "share": MessageLookupByLibrary.simpleMessage("Share"),
+        "signOut": MessageLookupByLibrary.simpleMessage("Sign out"),
+        "skip": MessageLookupByLibrary.simpleMessage("Skip"),
+        "smallLessThan20Employee":
+            MessageLookupByLibrary.simpleMessage("1 - 20 Employee"),
+        "startWriting": MessageLookupByLibrary.simpleMessage("Start Writing"),
+        "stcPayCode": MessageLookupByLibrary.simpleMessage("STC Pay Code"),
+        "store": MessageLookupByLibrary.simpleMessage("Store"),
+        "storeName": MessageLookupByLibrary.simpleMessage("Store Name"),
+        "storeOwner": MessageLookupByLibrary.simpleMessage("Store Owner"),
+        "storePhone": MessageLookupByLibrary.simpleMessage("Store Phone"),
+        "submittingProfile":
+            MessageLookupByLibrary.simpleMessage("Submitting Profile"),
+        "subscribe": MessageLookupByLibrary.simpleMessage("Subscribe"),
+        "successRenew": MessageLookupByLibrary.simpleMessage(
+            "Your subscription has been renewed"),
+        "sureForDelete": MessageLookupByLibrary.simpleMessage(
+            "Are you sure about deleting this order , please confirm"),
+        "syria": MessageLookupByLibrary.simpleMessage("Syria"),
+        "termsOfService":
+            MessageLookupByLibrary.simpleMessage("Terms of Service"),
+        "thisMightTakeAWhilePleaseWait": MessageLookupByLibrary.simpleMessage(
+            "This Might Take a while, please wait"),
+        "to": MessageLookupByLibrary.simpleMessage("To"),
+        "toFindOutMorePleaseLeaveYourPhonenandWeWill":
+            MessageLookupByLibrary.simpleMessage(
+                "To find out more, please leave your phone.\nand we will contact you personally"),
+        "took": MessageLookupByLibrary.simpleMessage("Took"),
+        "unaccept": MessageLookupByLibrary.simpleMessage(
+            "Your request not accepted from administration you can contact us to get info"),
+        "unknown": MessageLookupByLibrary.simpleMessage("Unknown"),
+        "unknownNumberOfCar":
+            MessageLookupByLibrary.simpleMessage("The number of cars unknown"),
+        "updateBranch": MessageLookupByLibrary.simpleMessage("Update Branch"),
+        "updateBranchFailure": MessageLookupByLibrary.simpleMessage(
+            "Sorry your request for update branch faild please try again"),
+        "updateBranchSuccess": MessageLookupByLibrary.simpleMessage(
+            "Branch has been updated successfully"),
+        "updateBranches":
+            MessageLookupByLibrary.simpleMessage("Update Branches"),
+        "uploadAndSubmit":
+            MessageLookupByLibrary.simpleMessage("Upload and submit"),
+        "uploadingAndSubmitting":
+            MessageLookupByLibrary.simpleMessage("Uploading and Submitting"),
+        "uploadingImages":
+            MessageLookupByLibrary.simpleMessage("Uploading Images"),
+        "uploadingImagesPleaseWait": MessageLookupByLibrary.simpleMessage(
+            "Uploading Images, Please Wait"),
+        "warnning": MessageLookupByLibrary.simpleMessage("Important Note"),
+        "weDeliver": MessageLookupByLibrary.simpleMessage("We Deliver"),
+        "weNeedYourImageAndYourDriverLicenceInOrder":
+            MessageLookupByLibrary.simpleMessage(
+                "We need your image and your driver licence in order to sign up"),
+        "weNeedYourImageInOrderToSignUp": MessageLookupByLibrary.simpleMessage(
+            "We need  your image in order to sign up"),
+        "weWillContactYouSoon":
+            MessageLookupByLibrary.simpleMessage("We will contact you soon"),
+        "welcomeToOrdersScreen":
+            MessageLookupByLibrary.simpleMessage("Welcome to Orders Screen"),
+        "welcomeTopasco_shipping": MessageLookupByLibrary.simpleMessage(
+            "Welcome to pasco_shipping"),
+        "whatsappWithCaptain": MessageLookupByLibrary.simpleMessage("Captain"),
+        "whatsappWithClient": MessageLookupByLibrary.simpleMessage("Client"),
+        "whatsappWithStoreOwner":
+            MessageLookupByLibrary.simpleMessage("Store Owner")
+      };
 }
