@@ -10,28 +10,43 @@ import 'intl/messages_all.dart';
 
 // ignore_for_file: non_constant_identifier_names, lines_longer_than_80_chars
 // ignore_for_file: join_return_with_assignment, prefer_final_in_for_each
-// ignore_for_file: avoid_redundant_argument_values
+// ignore_for_file: avoid_redundant_argument_values, avoid_escaping_inner_quotes
 
 class S {
   S();
-  
-  static S current;
-  
-  static const AppLocalizationDelegate delegate =
-    AppLocalizationDelegate();
+
+  static S? _current;
+
+  static S get current {
+    assert(_current != null,
+        'No instance of S was loaded. Try to initialize the S delegate before accessing S.current.');
+    return _current!;
+  }
+
+  static const AppLocalizationDelegate delegate = AppLocalizationDelegate();
 
   static Future<S> load(Locale locale) {
-    final name = (locale.countryCode?.isEmpty ?? false) ? locale.languageCode : locale.toString();
-    final localeName = Intl.canonicalizedLocale(name); 
+    final name = (locale.countryCode?.isEmpty ?? false)
+        ? locale.languageCode
+        : locale.toString();
+    final localeName = Intl.canonicalizedLocale(name);
     return initializeMessages(localeName).then((_) {
       Intl.defaultLocale = localeName;
-      S.current = S();
-      
-      return S.current;
+      final instance = S();
+      S._current = instance;
+
+      return instance;
     });
-  } 
+  }
 
   static S of(BuildContext context) {
+    final instance = S.maybeOf(context);
+    assert(instance != null,
+        'No instance of S present in the widget tree. Did you add S.delegate in localizationsDelegates?');
+    return instance!;
+  }
+
+  static S? maybeOf(BuildContext context) {
     return Localizations.of<S>(context, S);
   }
 
@@ -384,6 +399,1866 @@ class S {
       args: [],
     );
   }
+
+  /// `Order Created, Return to Orders`
+  String get orderCreatedReturnToOrders {
+    return Intl.message(
+      'Order Created, Return to Orders',
+      name: 'orderCreatedReturnToOrders',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `New Order`
+  String get newOrder {
+    return Intl.message(
+      'New Order',
+      name: 'newOrder',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Renew Subscription`
+  String get renewSubscription {
+    return Intl.message(
+      'Renew Subscription',
+      name: 'renewSubscription',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Register Success, Setup my profile`
+  String get registerSuccessSetupMyProfile {
+    return Intl.message(
+      'Register Success, Setup my profile',
+      name: 'registerSuccessSetupMyProfile',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Choose Your Size`
+  String get chooseYourSize {
+    return Intl.message(
+      'Choose Your Size',
+      name: 'chooseYourSize',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `1 - 20 Employee`
+  String get smallLessThan20Employee {
+    return Intl.message(
+      '1 - 20 Employee',
+      name: 'smallLessThan20Employee',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `21 - 100 Employees`
+  String get mediumMoreThan20EmployeesLessThan100 {
+    return Intl.message(
+      '21 - 100 Employees',
+      name: 'mediumMoreThan20EmployeesLessThan100',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `+100 Employees`
+  String get largeMoreThan100Employees {
+    return Intl.message(
+      '+100 Employees',
+      name: 'largeMoreThan100Employees',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Open Chat Room`
+  String get openChatRoom {
+    return Intl.message(
+      'Open Chat Room',
+      name: 'openChatRoom',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Orders`
+  String get orders {
+    return Intl.message(
+      'Orders',
+      name: 'orders',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Current Orders`
+  String get currentOrders {
+    return Intl.message(
+      'Current Orders',
+      name: 'currentOrders',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Nearby Orders`
+  String get nearbyOrders {
+    return Intl.message(
+      'Nearby Orders',
+      name: 'nearbyOrders',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Chat with Store Owner`
+  String get chatWithStoreOwner {
+    return Intl.message(
+      'Chat with Store Owner',
+      name: 'chatWithStoreOwner',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Store Owner`
+  String get whatsappWithStoreOwner {
+    return Intl.message(
+      'Store Owner',
+      name: 'whatsappWithStoreOwner',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Client`
+  String get whatsappWithClient {
+    return Intl.message(
+      'Client',
+      name: 'whatsappWithClient',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Get Direction`
+  String get getDirection {
+    return Intl.message(
+      'Get Direction',
+      name: 'getDirection',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Settings`
+  String get settings {
+    return Intl.message(
+      'Settings',
+      name: 'settings',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `I am`
+  String get iAm {
+    return Intl.message(
+      'I am',
+      name: 'iAm',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Skip`
+  String get skip {
+    return Intl.message(
+      'Skip',
+      name: 'skip',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Preferred language`
+  String get iSpeak {
+    return Intl.message(
+      'Preferred language',
+      name: 'iSpeak',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Change Language`
+  String get changeLanguage {
+    return Intl.message(
+      'Change Language',
+      name: 'changeLanguage',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `And I am`
+  String get andIAm {
+    return Intl.message(
+      'And I am',
+      name: 'andIAm',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `ME`
+  String get me {
+    return Intl.message(
+      'ME',
+      name: 'me',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `We need your image and your driver licence in order to sign up`
+  String get weNeedYourImageAndYourDriverLicenceInOrder {
+    return Intl.message(
+      'We need your image and your driver licence in order to sign up',
+      name: 'weNeedYourImageAndYourDriverLicenceInOrder',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `We need  your image in order to sign up`
+  String get weNeedYourImageInOrderToSignUp {
+    return Intl.message(
+      'We need  your image in order to sign up',
+      name: 'weNeedYourImageInOrderToSignUp',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Upload and submit`
+  String get uploadAndSubmit {
+    return Intl.message(
+      'Upload and submit',
+      name: 'uploadAndSubmit',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `This Might Take a while, please wait`
+  String get thisMightTakeAWhilePleaseWait {
+    return Intl.message(
+      'This Might Take a while, please wait',
+      name: 'thisMightTakeAWhilePleaseWait',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Uploading Images, Please Wait`
+  String get uploadingImagesPleaseWait {
+    return Intl.message(
+      'Uploading Images, Please Wait',
+      name: 'uploadingImagesPleaseWait',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Driver Licence`
+  String get driverLicence {
+    return Intl.message(
+      'Driver Licence',
+      name: 'driverLicence',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Age`
+  String get age {
+    return Intl.message(
+      'Age',
+      name: 'age',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Uploading and Submitting`
+  String get uploadingAndSubmitting {
+    return Intl.message(
+      'Uploading and Submitting',
+      name: 'uploadingAndSubmitting',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Uploading Images`
+  String get uploadingImages {
+    return Intl.message(
+      'Uploading Images',
+      name: 'uploadingImages',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Submitting Profile`
+  String get submittingProfile {
+    return Intl.message(
+      'Submitting Profile',
+      name: 'submittingProfile',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Error Uploading Images!!`
+  String get errorUploadingImages {
+    return Intl.message(
+      'Error Uploading Images!!',
+      name: 'errorUploadingImages',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Account Created`
+  String get accountCreated {
+    return Intl.message(
+      'Account Created',
+      name: 'accountCreated',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Move to Orders`
+  String get moveToOrders {
+    return Intl.message(
+      'Move to Orders',
+      name: 'moveToOrders',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Contact Phone Number`
+  String get contactPhoneNumber {
+    return Intl.message(
+      'Contact Phone Number',
+      name: 'contactPhoneNumber',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Pay Subscription`
+  String get paySubscription {
+    return Intl.message(
+      'Pay Subscription',
+      name: 'paySubscription',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Earn Cash`
+  String get earnCash {
+    return Intl.message(
+      'Earn Cash',
+      name: 'earnCash',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Deliver`
+  String get deliver {
+    return Intl.message(
+      'Deliver',
+      name: 'deliver',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Accept`
+  String get accept {
+    return Intl.message(
+      'Accept',
+      name: 'accept',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Check Orders`
+  String get checkOrders {
+    return Intl.message(
+      'Check Orders',
+      name: 'checkOrders',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Open the app`
+  String get openTheApp {
+    return Intl.message(
+      'Open the app',
+      name: 'openTheApp',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Book a car`
+  String get bookACar {
+    return Intl.message(
+      'Book a car',
+      name: 'bookACar',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `We Deliver`
+  String get weDeliver {
+    return Intl.message(
+      'We Deliver',
+      name: 'weDeliver',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Help me more`
+  String get helpMeMore {
+    return Intl.message(
+      'Help me more',
+      name: 'helpMeMore',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Request Meeting`
+  String get requestMeeting {
+    return Intl.message(
+      'Request Meeting',
+      name: 'requestMeeting',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `To find out more, please leave your phone.\nand we will contact you personally`
+  String get toFindOutMorePleaseLeaveYourPhonenandWeWill {
+    return Intl.message(
+      'To find out more, please leave your phone.\nand we will contact you personally',
+      name: 'toFindOutMorePleaseLeaveYourPhonenandWeWill',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `We will contact you soon`
+  String get weWillContactYouSoon {
+    return Intl.message(
+      'We will contact you soon',
+      name: 'weWillContactYouSoon',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Welcome to pasco_shipping`
+  String get welcomeTopasco_shipping {
+    return Intl.message(
+      'Welcome to pasco_shipping',
+      name: 'welcomeTopasco_shipping',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Init Data`
+  String get initData {
+    return Intl.message(
+      'Init Data',
+      name: 'initData',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Save Location as Branch 01`
+  String get saveLocationAsBranch01 {
+    return Intl.message(
+      'Save Location as Branch 01',
+      name: 'saveLocationAsBranch01',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Welcome to Orders Screen`
+  String get welcomeToOrdersScreen {
+    return Intl.message(
+      'Welcome to Orders Screen',
+      name: 'welcomeToOrdersScreen',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Profile`
+  String get profile {
+    return Intl.message(
+      'Profile',
+      name: 'profile',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Email`
+  String get email {
+    return Intl.message(
+      'Email',
+      name: 'email',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Password is too short`
+  String get passwordIsTooShort {
+    return Intl.message(
+      'Password is too short',
+      name: 'passwordIsTooShort',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Password`
+  String get password {
+    return Intl.message(
+      'Password',
+      name: 'password',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Name`
+  String get name {
+    return Intl.message(
+      'Name',
+      name: 'name',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Name is required`
+  String get nameIsRequired {
+    return Intl.message(
+      'Name is required',
+      name: 'nameIsRequired',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Choose Your City`
+  String get chooseYourCity {
+    return Intl.message(
+      'Choose Your City',
+      name: 'chooseYourCity',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Next`
+  String get next {
+    return Intl.message(
+      'Next',
+      name: 'next',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `From WhatsApp`
+  String get fromWhatsapp {
+    return Intl.message(
+      'From WhatsApp',
+      name: 'fromWhatsapp',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Default Branch`
+  String get defaultBranch {
+    return Intl.message(
+      'Default Branch',
+      name: 'defaultBranch',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Cash`
+  String get cash {
+    return Intl.message(
+      'Cash',
+      name: 'cash',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `online`
+  String get online {
+    return Intl.message(
+      'online',
+      name: 'online',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `To`
+  String get to {
+    return Intl.message(
+      'To',
+      name: 'to',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Info`
+  String get info {
+    return Intl.message(
+      'Info',
+      name: 'info',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Mohammad`
+  String get mohammad {
+    return Intl.message(
+      'Mohammad',
+      name: 'mohammad',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Deliver To`
+  String get deliverTo {
+    return Intl.message(
+      'Deliver To',
+      name: 'deliverTo',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Captain`
+  String get whatsappWithCaptain {
+    return Intl.message(
+      'Captain',
+      name: 'whatsappWithCaptain',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Order`
+  String get order {
+    return Intl.message(
+      'Order',
+      name: 'order',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Home`
+  String get home {
+    return Intl.message(
+      'Home',
+      name: 'home',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Branch 01`
+  String get branch01 {
+    return Intl.message(
+      'Branch 01',
+      name: 'branch01',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Create new order`
+  String get createNewOrder {
+    return Intl.message(
+      'Create new order',
+      name: 'createNewOrder',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Register`
+  String get register {
+    return Intl.message(
+      'Register',
+      name: 'register',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Branch`
+  String get branch {
+    return Intl.message(
+      'Branch',
+      name: 'branch',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Save Branches`
+  String get saveBranches {
+    return Intl.message(
+      'Save Branches',
+      name: 'saveBranches',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Bank Name`
+  String get bankName {
+    return Intl.message(
+      'Bank Name',
+      name: 'bankName',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Account Number`
+  String get accountNumber {
+    return Intl.message(
+      'Account Number',
+      name: 'accountNumber',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Save`
+  String get save {
+    return Intl.message(
+      'Save',
+      name: 'save',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Dark Mode`
+  String get darkMode {
+    return Intl.message(
+      'Dark Mode',
+      name: 'darkMode',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Error Loading Branches`
+  String get errorLoadingBranches {
+    return Intl.message(
+      'Error Loading Branches',
+      name: 'errorLoadingBranches',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `CANCEL`
+  String get cancel {
+    return Intl.message(
+      'CANCEL',
+      name: 'cancel',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Our Bank Name`
+  String get ourBankName {
+    return Intl.message(
+      'Our Bank Name',
+      name: 'ourBankName',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Our Bank Account Number`
+  String get ourBankAccountNumber {
+    return Intl.message(
+      'Our Bank Account Number',
+      name: 'ourBankAccountNumber',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Package`
+  String get package {
+    return Intl.message(
+      'Package',
+      name: 'package',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Order/Month`
+  String get ordermonth {
+    return Intl.message(
+      'Order/Month',
+      name: 'ordermonth',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Car`
+  String get car {
+    return Intl.message(
+      'Car',
+      name: 'car',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `APPLY`
+  String get apply {
+    return Intl.message(
+      'APPLY',
+      name: 'apply',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Recipient Phone Number`
+  String get recipientPhoneNumber {
+    return Intl.message(
+      'Recipient Phone Number',
+      name: 'recipientPhoneNumber',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Pick a Language`
+  String get pickALanguage {
+    return Intl.message(
+      'Pick a Language',
+      name: 'pickALanguage',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Pick a Job`
+  String get pickAJob {
+    return Intl.message(
+      'Pick a Job',
+      name: 'pickAJob',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Payment Method`
+  String get paymentMethod {
+    return Intl.message(
+      'Payment Method',
+      name: 'paymentMethod',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Please Select a Branch`
+  String get pleaseSelectABranch {
+    return Intl.message(
+      'Please Select a Branch',
+      name: 'pleaseSelectABranch',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Privacy Policy`
+  String get privacyPolicy {
+    return Intl.message(
+      'Privacy Policy',
+      name: 'privacyPolicy',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Terms of Service`
+  String get termsOfService {
+    return Intl.message(
+      'Terms of Service',
+      name: 'termsOfService',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Created`
+  String get created {
+    return Intl.message(
+      'Created',
+      name: 'created',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Unknown`
+  String get unknown {
+    return Intl.message(
+      'Unknown',
+      name: 'unknown',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Captain Accepted Order`
+  String get captainAcceptedOrder {
+    return Intl.message(
+      'Captain Accepted Order',
+      name: 'captainAcceptedOrder',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Captain in Store`
+  String get captainInStore {
+    return Intl.message(
+      'Captain in Store',
+      name: 'captainInStore',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Captain Started Delivery`
+  String get captainStartedDelivery {
+    return Intl.message(
+      'Captain Started Delivery',
+      name: 'captainStartedDelivery',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Captain Got Cash`
+  String get captainGotCash {
+    return Intl.message(
+      'Captain Got Cash',
+      name: 'captainGotCash',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Order is finished`
+  String get orderIsFinished {
+    return Intl.message(
+      'Order is finished',
+      name: 'orderIsFinished',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Order is created`
+  String get orderIsCreated {
+    return Intl.message(
+      'Order is created',
+      name: 'orderIsCreated',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Activity Log`
+  String get activityLog {
+    return Intl.message(
+      'Activity Log',
+      name: 'activityLog',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Save Success`
+  String get saveSuccess {
+    return Intl.message(
+      'Save Success',
+      name: 'saveSuccess',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `My Profile`
+  String get myProfile {
+    return Intl.message(
+      'My Profile',
+      name: 'myProfile',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Reason of the report`
+  String get reasonOfTheReport {
+    return Intl.message(
+      'Reason of the report',
+      name: 'reasonOfTheReport',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Create new Report`
+  String get createNewReport {
+    return Intl.message(
+      'Create new Report',
+      name: 'createNewReport',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Help`
+  String get help {
+    return Intl.message(
+      'Help',
+      name: 'help',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Please Complete the Form`
+  String get pleaseCompleteTheForm {
+    return Intl.message(
+      'Please Complete the Form',
+      name: 'pleaseCompleteTheForm',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Report sent`
+  String get reportSent {
+    return Intl.message(
+      'Report sent',
+      name: 'reportSent',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Reason is Required`
+  String get reasonIsRequired {
+    return Intl.message(
+      'Reason is Required',
+      name: 'reasonIsRequired',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Please Provide us with the Client Name`
+  String get pleaseProvideUsWithTheClientName {
+    return Intl.message(
+      'Please Provide us with the Client Name',
+      name: 'pleaseProvideUsWithTheClientName',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Please Provide us the client phone number`
+  String get pleaseProvideUsTheClientPhoneNumber {
+    return Intl.message(
+      'Please Provide us the client phone number',
+      name: 'pleaseProvideUsTheClientPhoneNumber',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Store Name`
+  String get storeName {
+    return Intl.message(
+      'Store Name',
+      name: 'storeName',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Store Phone`
+  String get storePhone {
+    return Intl.message(
+      'Store Phone',
+      name: 'storePhone',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Phone is required`
+  String get phoneIsRequired {
+    return Intl.message(
+      'Phone is required',
+      name: 'phoneIsRequired',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `New name`
+  String get newName {
+    return Intl.message(
+      'New name',
+      name: 'newName',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Our Packages`
+  String get ourPackages {
+    return Intl.message(
+      'Our Packages',
+      name: 'ourPackages',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Notifications`
+  String get notifications {
+    return Intl.message(
+      'Notifications',
+      name: 'notifications',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Please Provide the distance`
+  String get pleaseProvideTheDistance {
+    return Intl.message(
+      'Please Provide the distance',
+      name: 'pleaseProvideTheDistance',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Finish Order, Provide distance in KM`
+  String get finishOrderProvideDistanceInKm {
+    return Intl.message(
+      'Finish Order, Provide distance in KM',
+      name: 'finishOrderProvideDistanceInKm',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `My Status`
+  String get myStatus {
+    return Intl.message(
+      'My Status',
+      name: 'myStatus',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Direct Support`
+  String get directSupport {
+    return Intl.message(
+      'Direct Support',
+      name: 'directSupport',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `My Orders`
+  String get myOrders {
+    return Intl.message(
+      'My Orders',
+      name: 'myOrders',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Share`
+  String get share {
+    return Intl.message(
+      'Share',
+      name: 'share',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Please Download pasco_shipping`
+  String get pleaseDownloadpasco_shipping {
+    return Intl.message(
+      'Please Download pasco_shipping',
+      name: 'pleaseDownloadpasco_shipping',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `My Plan`
+  String get myPlan {
+    return Intl.message(
+      'My Plan',
+      name: 'myPlan',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Bank Account Number`
+  String get bankAccountNumber {
+    return Intl.message(
+      'Bank Account Number',
+      name: 'bankAccountNumber',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `STC Pay Code`
+  String get stcPayCode {
+    return Intl.message(
+      'STC Pay Code',
+      name: 'stcPayCode',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Current Balance:`
+  String get currentBalance {
+    return Intl.message(
+      'Current Balance:',
+      name: 'currentBalance',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Next Payment Date:`
+  String get nextPaymentDate {
+    return Intl.message(
+      'Next Payment Date:',
+      name: 'nextPaymentDate',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Saudi Riyal`
+  String get saudiRiyal {
+    return Intl.message(
+      'Saudi Riyal',
+      name: 'saudiRiyal',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Active Plan`
+  String get activePlan {
+    return Intl.message(
+      'Active Plan',
+      name: 'activePlan',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Active Cars`
+  String get activeCars {
+    return Intl.message(
+      'Active Cars',
+      name: 'activeCars',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Orders / Month`
+  String get ordersMonth {
+    return Intl.message(
+      'Orders / Month',
+      name: 'ordersMonth',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Payment History`
+  String get paymentHistory {
+    return Intl.message(
+      'Payment History',
+      name: 'paymentHistory',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Completed Orders`
+  String get completedOrders {
+    return Intl.message(
+      'Completed Orders',
+      name: 'completedOrders',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Latest Updates`
+  String get latestUpdates {
+    return Intl.message(
+      'Latest Updates',
+      name: 'latestUpdates',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `My Balance`
+  String get myBalance {
+    return Intl.message(
+      'My Balance',
+      name: 'myBalance',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `No Image`
+  String get noImage {
+    return Intl.message(
+      'No Image',
+      name: 'noImage',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `How we work`
+  String get howWeWork {
+    return Intl.message(
+      'How we work',
+      name: 'howWeWork',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Took`
+  String get took {
+    return Intl.message(
+      'Took',
+      name: 'took',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `minute`
+  String get minutes {
+    return Intl.message(
+      'minute',
+      name: 'minutes',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `I Agree to the Terms os service.`
+  String get iAgreeToTheTermsOsService {
+    return Intl.message(
+      'I Agree to the Terms os service.',
+      name: 'iAgreeToTheTermsOsService',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `I Agree To The Terms Of Service & privacy policy`
+  String get iAgreeToTheTermsOfServicePrivacyPolicy {
+    return Intl.message(
+      'I Agree To The Terms Of Service & privacy policy',
+      name: 'iAgreeToTheTermsOfServicePrivacyPolicy',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Business Name`
+  String get businessName {
+    return Intl.message(
+      'Business Name',
+      name: 'businessName',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `The number of cars unknown`
+  String get unknownNumberOfCar {
+    return Intl.message(
+      'The number of cars unknown',
+      name: 'unknownNumberOfCar',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Recipient Name`
+  String get recipientName {
+    return Intl.message(
+      'Recipient Name',
+      name: 'recipientName',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Date`
+  String get date {
+    return Intl.message(
+      'Date',
+      name: 'date',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Go`
+  String get go {
+    return Intl.message(
+      'Go',
+      name: 'go',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Error logging in, firebase account not found`
+  String get errorLoggingInFirebaseAccountNotFound {
+    return Intl.message(
+      'Error logging in, firebase account not found',
+      name: 'errorLoggingInFirebaseAccountNotFound',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Location of customer`
+  String get locationOfCustomer {
+    return Intl.message(
+      'Location of customer',
+      name: 'locationOfCustomer',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Are you sure about deleting this order , please confirm`
+  String get sureForDelete {
+    return Intl.message(
+      'Are you sure about deleting this order , please confirm',
+      name: 'sureForDelete',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `the order has been deleted successfully`
+  String get deleteSuccess {
+    return Intl.message(
+      'the order has been deleted successfully',
+      name: 'deleteSuccess',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `distance in KM`
+  String get ProvideDistanceInKm {
+    return Intl.message(
+      'distance in KM',
+      name: 'ProvideDistanceInKm',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Your subscription not verified yet`
+  String get notVerified {
+    return Intl.message(
+      'Your subscription not verified yet',
+      name: 'notVerified',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Subscription options`
+  String get renewPlan {
+    return Intl.message(
+      'Subscription options',
+      name: 'renewPlan',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Renew old package`
+  String get renewOldPlan {
+    return Intl.message(
+      'Renew old package',
+      name: 'renewOldPlan',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Subscription with new package`
+  String get renewNewPlan {
+    return Intl.message(
+      'Subscription with new package',
+      name: 'renewNewPlan',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Your subscription has been renewed`
+  String get successRenew {
+    return Intl.message(
+      'Your subscription has been renewed',
+      name: 'successRenew',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Important Note`
+  String get warnning {
+    return Intl.message(
+      'Important Note',
+      name: 'warnning',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Your subscription not activate yet please wait`
+  String get inactive {
+    return Intl.message(
+      'Your subscription not activate yet please wait',
+      name: 'inactive',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `You reach your limit of orders you can renew your subscription from settings`
+  String get outOforders {
+    return Intl.message(
+      'You reach your limit of orders you can renew your subscription from settings',
+      name: 'outOforders',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Your subscription is out of date please renew your subscription`
+  String get finishedDate {
+    return Intl.message(
+      'Your subscription is out of date please renew your subscription',
+      name: 'finishedDate',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Your request not accepted from administration you can contact us to get info`
+  String get unaccept {
+    return Intl.message(
+      'Your request not accepted from administration you can contact us to get info',
+      name: 'unaccept',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `There is no car available please wait your ongoing orders to finish`
+  String get outOfCars {
+    return Intl.message(
+      'There is no car available please wait your ongoing orders to finish',
+      name: 'outOfCars',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `You dont have a subscription yet`
+  String get notSubscription {
+    return Intl.message(
+      'You dont have a subscription yet',
+      name: 'notSubscription',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Subscribe`
+  String get subscribe {
+    return Intl.message(
+      'Subscribe',
+      name: 'subscribe',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Your account not activated yet please wait`
+  String get captainNotActive {
+    return Intl.message(
+      'Your account not activated yet please wait',
+      name: 'captainNotActive',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `day`
+  String get days {
+    return Intl.message(
+      'day',
+      name: 'days',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `second`
+  String get seconds {
+    return Intl.message(
+      'second',
+      name: 'seconds',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `hour`
+  String get hours {
+    return Intl.message(
+      'hour',
+      name: 'hours',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Your renew has been failed please try again`
+  String get failureRenew {
+    return Intl.message(
+      'Your renew has been failed please try again',
+      name: 'failureRenew',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Store`
+  String get store {
+    return Intl.message(
+      'Store',
+      name: 'store',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Update Branches`
+  String get updateBranches {
+    return Intl.message(
+      'Update Branches',
+      name: 'updateBranches',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Update Branch`
+  String get updateBranch {
+    return Intl.message(
+      'Update Branch',
+      name: 'updateBranch',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Save Branch`
+  String get saveBranch {
+    return Intl.message(
+      'Save Branch',
+      name: 'saveBranch',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Edit Branch Name`
+  String get editBranchName {
+    return Intl.message(
+      'Edit Branch Name',
+      name: 'editBranchName',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Branch has been updated successfully`
+  String get updateBranchSuccess {
+    return Intl.message(
+      'Branch has been updated successfully',
+      name: 'updateBranchSuccess',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Sorry your request for update branch faild please try again`
+  String get updateBranchFailure {
+    return Intl.message(
+      'Sorry your request for update branch faild please try again',
+      name: 'updateBranchFailure',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Phone number is too short`
+  String get phoneNumbertooShort {
+    return Intl.message(
+      'Phone number is too short',
+      name: 'phoneNumbertooShort',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Add Branch`
+  String get addBranch {
+    return Intl.message(
+      'Add Branch',
+      name: 'addBranch',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Branch has been added successfully`
+  String get addBranchSuccess {
+    return Intl.message(
+      'Branch has been added successfully',
+      name: 'addBranchSuccess',
+      desc: '',
+      args: [],
+    );
+  }
+
+  /// `Sorry your request for add branch faild please try again`
+  String get addBranchFailure {
+    return Intl.message(
+      'Sorry your request for add branch faild please try again',
+      name: 'addBranchFailure',
+      desc: '',
+      args: [],
+    );
+  }
 }
 
 class AppLocalizationDelegate extends LocalizationsDelegate<S> {
@@ -404,11 +2279,9 @@ class AppLocalizationDelegate extends LocalizationsDelegate<S> {
   bool shouldReload(AppLocalizationDelegate old) => false;
 
   bool _isSupported(Locale locale) {
-    if (locale != null) {
-      for (var supportedLocale in supportedLocales) {
-        if (supportedLocale.languageCode == locale.languageCode) {
-          return true;
-        }
+    for (var supportedLocale in supportedLocales) {
+      if (supportedLocale.languageCode == locale.languageCode) {
+        return true;
       }
     }
     return false;
