@@ -1,20 +1,22 @@
-import 'package:yessoft/abstracts/module/yes_module.dart';
+import 'package:injectable/injectable.dart';
+import 'package:pasco_shipping/abstracts/module/yes_module.dart';
+import 'package:pasco_shipping/module_auth/ui/screen/login_screen/login_screen.dart';
+import 'package:pasco_shipping/module_auth/ui/screen/register_screen/register_screen.dart';
 import 'package:flutter/cupertino.dart';
-import 'package:flutter/material.dart';
-import 'package:inject/inject.dart';
-
 import 'authorization_routes.dart';
-import 'ui/screen/login_screen/auth_screen.dart';
 
-@provide
+@injectable
 class AuthorizationModule extends YesModule {
-  final AuthScreen _authScreen;
-  AuthorizationModule(this._authScreen);
+  final LoginScreen _loginScreen;
+  final RegisterScreen _registerScreen;
 
-  @override
+  AuthorizationModule(this._loginScreen, this._registerScreen) {
+    YesModule.RoutesMap.addAll(getRoutes());
+  }
   Map<String, WidgetBuilder> getRoutes() {
     return {
-      AuthorizationRoutes.AUTH_SCREEN: (context) => _authScreen,
+      AuthorizationRoutes.LOGIN_SCREEN: (context) => _loginScreen,
+      AuthorizationRoutes.REGISTER_SCREEN: (context) => _registerScreen,
     };
   }
 }
