@@ -1,13 +1,23 @@
-import 'package:yessoft/generated/l10n.dart';
+import 'package:pasco_shipping/generated/l10n.dart';
 import 'package:flutter/material.dart';
 
 class PhoneLoginWidget extends StatefulWidget {
-  final Function(String) onLoginRequested;
+  final Function(String)? onLoginRequested;
+  final Function()? onRetry;
+  final Function(String)? onConfirm;
+
+  // To switch between register and login
+  final Function()? onAlterRequest;
+  final bool? codeSent;
   final bool isRegister;
 
   PhoneLoginWidget({
     this.onLoginRequested,
+    this.onConfirm,
+    this.onRetry,
+    this.codeSent,
     this.isRegister = false,
+    this.onAlterRequest,
   });
 
   @override
@@ -15,24 +25,46 @@ class PhoneLoginWidget extends StatefulWidget {
 }
 
 class _PhoneLoginWidgetState extends State<PhoneLoginWidget> {
+  String? _errorMsg;
+  Scaffold? pageLayout;
+  bool loading = false;
+
   final GlobalKey _signUpFormKey = GlobalKey<FormState>();
   final TextEditingController _phoneController = TextEditingController();
-
   String countryCode = '+963';
+
+  bool retryEnabled = false;
+  bool agreed = false;
+  String hint = '900000000';
 
   @override
   Widget build(BuildContext context) {
+    if (loading) {
+      Future.delayed(Duration(seconds: 30)).then((value) {
+        loading = false;
+        if (mounted) {
+          setState(() {});
+        }
+      });
+    }
+    if (countryCode == '+963') {
+      hint = '900000000';
+    } else if (countryCode == '+966') {
+      hint = '500000000';
+    } else {
+      hint = '700000000';
+    }
     return SingleChildScrollView(
       child: Form(
         key: _signUpFormKey,
         child: Flex(
           direction: Axis.vertical,
-          mainAxisAlignment: MainAxisAlignment.center,
+          mainAxisAlignment: MainAxisAlignment.spaceAround,
           children: [
-            Padding(
-              padding: const EdgeInsets.all(24.0),
-              child: FlutterLogo(),
-            ),
+            MediaQuery.of(context).viewInsets.bottom == 0
+                ? Container(
+                    height: 144, child: Image.asset('assets/images/track.png'))
+                : Container(),
             Padding(
               padding: const EdgeInsets.all(16.0),
               child: Row(
@@ -40,7 +72,7 @@ class _PhoneLoginWidgetState extends State<PhoneLoginWidget> {
                 children: [
                   DropdownButton(
                     onChanged: (v) {
-                      countryCode = v;
+                      countryCode = v.toString();
                       if (mounted) setState(() {});
                     },
                     value: countryCode,
@@ -66,9 +98,9 @@ class _PhoneLoginWidgetState extends State<PhoneLoginWidget> {
                         controller: _phoneController,
                         decoration: InputDecoration(
                             labelText: S.of(context).phoneNumber,
-                            hintText: '111 111 11111'),
+                            hintText: '$hint'),
                         validator: (v) {
-                          if (v.isEmpty) {
+                          if (v!.isEmpty) {
                             return S.of(context).pleaseInputPhoneNumber;
                           }
                           return null;
@@ -80,29 +112,59 @@ class _PhoneLoginWidgetState extends State<PhoneLoginWidget> {
                 ],
               ),
             ),
-            Container(
-              child: FlatButton(
-                color: widget.isRegister
-                    ? Theme.of(context).primaryColorDark
-                    : Theme.of(context).primaryColor,
-                onPressed: () {
-                  String phone = _phoneController.text;
-                  if (phone[0] == '0') {
-                    phone = phone.substring(1);
-                  }
-                  widget.onLoginRequested(countryCode + phone);
-                },
-                child: Padding(
-                  padding: const EdgeInsets.all(16.0),
-                  child: Text(
-                    S.of(context).sendMeCode,
-                    style: TextStyle(
-                      color: Colors.white,
-                    ),
-                  ),
-                ),
+            _errorMsg != null ? Text(_errorMsg!) : Container(),
+            GestureDetector(
+              onTap: () {
+                if (widget.onAlterRequest != null) widget.onAlterRequest!();
+              },
+              child: Padding(
+                padding: const EdgeInsets.all(16.0),
+                child: Text(widget.isRegister
+                    ? S.of(context).iHaveAnAccount
+                    : S.of(context).register),
               ),
-            )
+            ),
+            if (widget.isRegister != false)
+              CheckboxListTile(
+                  value: agreed,
+                  title: Text(S.of(context).iAgreeToTheTermsOsService),
+                  onChanged: (v) {
+                    agreed = v!;
+                    if (mounted) setState(() {});
+                  }),
+            loading == true
+                ? Center(
+                    child: Text(S.of(context).loading),
+                  )
+                : Container(
+                    child: FlatButton(
+                      shape: RoundedRectangleBorder(
+                          borderRadius: BorderRadius.circular(15)),
+                      color: widget.isRegister
+                          ? Theme.of(context).primaryColorDark
+                          : Theme.of(context).primaryColor,
+                      onPressed: !agreed && widget.isRegister
+                          ? null
+                          : () {
+                              String phone = _phoneController.text;
+                              if (phone[0] == '0') {
+                                phone = phone.substring(1);
+                              }
+                              loading = true;
+                              setState(() {});
+                              widget.onLoginRequested!(countryCode + phone);
+                            },
+                      child: Padding(
+                        padding: const EdgeInsets.all(16.0),
+                        child: Text(
+                          S.of(context).sendMeCode,
+                          style: TextStyle(
+                            color: Colors.white,
+                          ),
+                        ),
+                      ),
+                    ),
+                  )
           ],
         ),
       ),
