@@ -1,9 +1,9 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:inject/inject.dart';
-import 'package:yessoft/module_chat/model/chat/chat_model.dart';
-import 'package:yessoft/module_chat/repository/chat/chat_repository.dart';
+import 'package:injectable/injectable.dart';
+import 'package:pasco_shipping/module_chat/model/chat/chat_model.dart';
+import 'package:pasco_shipping/module_chat/repository/chat/chat_repository.dart';
 
-@provide
+@injectable
 class ChatManager {
   final ChatRepository _chatRepository;
 
@@ -15,5 +15,8 @@ class ChatManager {
 
   void sendMessage(String chatRoomID, ChatModel chatMessage) {
     _chatRepository.sendMessage(chatRoomID, chatMessage);
+  }
+  void sendNotification(String chatRoomID , bool support , bool feedBack){
+    _chatRepository.sendNotification(chatRoomID,support,feedBack);
   }
 }
