@@ -1,10 +1,10 @@
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:inject/inject.dart';
+import 'package:injectable/injectable.dart';
 import 'package:rxdart/rxdart.dart';
-import 'package:yessoft/module_chat/manager/chat/chat_manager.dart';
-import 'package:yessoft/module_chat/model/chat/chat_model.dart';
+import 'package:pasco_shipping/module_chat/manager/chat/chat_manager.dart';
+import 'package:pasco_shipping/module_chat/model/chat/chat_model.dart';
 
-@provide
+@injectable
 class ChatService {
   final ChatManager _chatManager;
 
@@ -12,7 +12,7 @@ class ChatService {
 
   // This is Real Time, That is Why I went this way
   final PublishSubject<List<ChatModel>> _chatPublishSubject =
-      new PublishSubject();
+  new PublishSubject();
 
   Stream<List<ChatModel>> get chatMessagesStream => _chatPublishSubject.stream;
 
@@ -20,22 +20,22 @@ class ChatService {
     _chatManager.getMessages(chatRoomID).listen((event) {
       List<ChatModel> chatMessagesList = [];
       event.docs.forEach((element) {
-        chatMessagesList.add(new ChatModel.fromJson(element.data()));
+        chatMessagesList.add(new ChatModel.fromJson(element.data() as Map<String,dynamic>));
       });
 
       _chatPublishSubject.add(chatMessagesList);
     });
   }
 
-  void sendMessage(String chatRoomID, String msg) async {
-    FirebaseAuth auth = await FirebaseAuth.instance;
-    User user = auth.currentUser;
+  void sendMessage(String chatRoomID, String msg,bool support,feedBack) async {
+    FirebaseAuth auth =  FirebaseAuth.instance;
+    User? user = auth.currentUser;
     ChatModel model = new ChatModel(
       msg: msg,
-      sender: user.uid,
-      sentDate: DateTime.now().toIso8601String(),
-    );
+      sender: user!.uid,
+      sentDate: DateTime.now().toString(),);
     _chatManager.sendMessage(chatRoomID, model);
+    _chatManager.sendNotification(chatRoomID,support ,feedBack);
   }
 
   void dispose() {
