@@ -1,17 +1,14 @@
-import 'package:yessoft/abstracts/module/yes_module.dart';
-import 'package:yessoft/module_settings/setting_routes.dart';
-import 'package:yessoft/module_settings/ui/settings_page/settings_page.dart';
-import 'package:flutter/material.dart';
-import 'package:inject/inject.dart';
+import 'package:injectable/injectable.dart';
+import 'package:pasco_shipping/abstracts/module/yes_module.dart';
+import 'package:pasco_shipping/module_settings/setting_routes.dart';
+import 'package:pasco_shipping/module_settings/ui/settings_page/settings_page.dart';
 
-@provide
+@injectable
 class SettingsModule extends YesModule {
   final SettingsScreen _settingsScreen;
-
-  SettingsModule(this._settingsScreen);
-
-  @override
-  Map<String, WidgetBuilder> getRoutes() {
-    return {SettingRoutes.ROUTE_SETTINGS: (context) => _settingsScreen};
+  SettingsModule(this._settingsScreen) {
+    YesModule.RoutesMap.addAll({
+      SettingRoutes.ROUTE_SETTINGS: (context) => _settingsScreen,
+    });
   }
 }
