@@ -1,7 +1,7 @@
-import 'package:inject/inject.dart';
 import 'package:firebase_crashlytics/firebase_crashlytics.dart';
+import 'package:injectable/injectable.dart';
 
-@provide
+@injectable
 @singleton
 class Logger {
   void info(String tag, String msg) {
@@ -14,11 +14,11 @@ class Logger {
     print('$time: \t $tag \t $msg');
   }
 
-  void error(String tag, String msg, StackTrace trace) {
+  void error(String tag, String msg, StackTrace? trace) {
     String time = DateTime.now().toString();
     print('$time: \t $tag \t $msg');
    FirebaseCrashlytics.instance
-       .recordError('$time: \t $tag \t $msg', trace??'');
+       .recordError('$time: \t $tag \t $msg', trace);
    FirebaseCrashlytics.instance.log('$time: \t $tag \t $msg');
    FirebaseCrashlytics.instance.sendUnsentReports();
   }
